### 2018-05-02 159664b
Minor fixes to JOSS paper
