### 2018-05-02 #712 @sjpfenninger
Noticed a broken bibtex ref in the article proof and [pushed an update to fix it](https://github.com/calliope-project/calliope/commit/159664ba3623ea0134ed473836f296b4485547c8)
