### 2018-05-02 0f59853
Updates to new storage constraints; added cyclic storage
