+++ calliope/test/test_backend_pyomo.py
@@ -173,7 +173,7 @@ class TestInterface:
         assert check_error_or_warning(error, 'Cannot rerun the backend in operate run mode')
 
 
-class TestConstraints:
+class TestBalanceConstraints:
 
     def test_loc_carriers_system_balance_constraint(self):
         """
@@ -233,6 +233,28 @@ class TestConstraints:
         m = build_model({}, 'simple_storage,two_hours,investment_costs')
         m.run(build_only=True)
         assert hasattr(m._backend_model, 'balance_storage_constraint')
+        assert not hasattr(m._backend_model, 'storage_initial_constraint')
+
+    def test_storage_initial_constraint(self):
+        """
+        sets.loc_techs_store,
+        """
+        m = build_model(
+            {'run.cyclic_storage': True},
+            'simple_storage,one_day,investment_costs'
+        )
+        m.run(build_only=True)
+        assert hasattr(m._backend_model, 'balance_storage_constraint')
+        assert not hasattr(m._backend_model, 'storage_initial_constraint')
+
+        m2 = build_model(
+            {'run.cyclic_storage': True,
+             'techs.test_storage.constraints.storage_initial': 0},
+            'simple_storage,one_day,investment_costs'
+        )
+        m2.run(build_only=True)
+        assert hasattr(m2._backend_model, 'balance_storage_constraint')
+        assert hasattr(m2._backend_model, 'storage_initial_constraint')
 
     def test_carriers_reserve_margin_constraint(self):
         """
@@ -242,6 +264,8 @@ class TestConstraints:
         m.run(build_only=True)
         assert hasattr(m._backend_model, 'reserve_margin_constraint')
 
+
+class TestCostConstraints:
     # costs.py
     def test_loc_techs_cost_constraint(self):
         """
@@ -315,6 +339,8 @@ class TestConstraints:
         m.run(build_only=True)
         assert hasattr(m._backend_model, 'cost_var_constraint')
 
+
+class TestExportConstraints:
     # export.py
     def test_loc_carriers_update_system_balance_constraint(self):
         """
@@ -384,6 +410,8 @@ class TestConstraints:
         m.run(build_only=True)
         assert hasattr(m._backend_model, 'export_max_constraint')
 
+
+class TestCapacityConstraints:
     # capacity.py
     def test_loc_techs_storage_capacity_constraint(self):
         """
@@ -654,6 +682,8 @@ class TestConstraints:
         m.run(build_only=True)
         assert hasattr(m._backend_model, 'energy_capacity_systemwide_constraint')
 
+
+class TestDispatchConstraints:
     # dispatch.py
     def test_loc_tech_carriers_carrier_production_max_constraint(self):
         """
@@ -668,7 +698,6 @@ class TestConstraints:
         m = build_model({}, 'supply_milp,two_hours,investment_costs')
         m.run(build_only=True)
         assert not hasattr(m._backend_model, 'carrier_production_max_constraint')
-
 
     def test_loc_tech_carriers_carrier_production_min_constraint(self):
         """
@@ -699,7 +728,6 @@ class TestConstraints:
         m.run(build_only=True)
         assert not hasattr(m._backend_model, 'carrier_production_min_constraint')
 
-
     def test_loc_tech_carriers_carrier_consumption_max_constraint(self):
         """
         i for i in sets.loc_tech_carriers_con
@@ -715,7 +743,6 @@ class TestConstraints:
         m = build_model({}, 'supply_milp,two_hours,investment_costs')
         m.run(build_only=True)
         assert hasattr(m._backend_model, 'carrier_consumption_max_constraint')
-
 
     def test_loc_techs_resource_max_constraint(self):
         """
@@ -778,6 +805,8 @@ class TestConstraints:
         assert hasattr(m._backend_model, 'ramping_up_constraint')
[…]
