### 2018-05-03 ef68089
Updates based on review
