### 2018-05-03 23f997c
AttrDict: fix union with all-empty subdicts
