### 2018-05-03 971bab9
Ensure pure dicts can be used to create calliope Model
