### 2018-05-03 c41c5bc
CLI: fix save_plots; tests: catch all warnings
