+++ calliope/cli.py
@@ -251,7 +251,7 @@ def run(model_file, override_file, save_netcdf, save_csv, save_plots,
         if save_plots:
             if termination == 'optimal':
                 print('Saving HTML file with plots to: {}'.format(save_plots))
-                model.plot.summary(out_file=save_plots)
+                model.plot.summary(to_file=save_plots)
             else:
                 click.secho(
                     'Model termination condition non-optimal. Not saving plots',
