### 2018-05-04 60894bb
Release v0.6.1
