+++ changelog.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+0.6.2-dev
+---------
+
+
 0.6.1 (2018-05-04)
 ------------------
 
