### 2018-05-04 91959ec
Tagline cleanup
