### 2018-05-04 73f76af
Fix empty iterables making it into model attributes
