### 2018-05-04 6aed7e8
Catch and kill cyclic storage in operate run mode
