+++ calliope/test/common/test_model/overrides.yaml
@@ -227,6 +227,13 @@ one_day:
 two_hours:
     model.subset_time: ['2005-01-01 00:00:00', '2005-01-01 01:00:00']
 
+operate:
+    run.mode: operate
+    model.subset_time: ['2005-01-01', '2005-01-04']
+    run.operation:
+            window: 12
+            horizon: 24
+
 investment_costs:
     tech_groups:
         supply:
