### 2018-05-04 1bc03e5
Fix CLI error when running minimal models
