+++ calliope/core/model.py
@@ -260,11 +260,11 @@ class Model(object):
         model_name = self._model_data.attrs.get('model.name', 'None')
         info_strings.append('Model name:   {}'.format(model_name))
         msize = '{locs} locations, {techs} technologies, {times} timesteps'.format(
-            locs=len(self._model_data.coords['locs'].values),
+            locs=len(self._model_data.coords.get('locs', [])),
             techs=(
-                len(self._model_data.coords['techs_non_transmission'].values) +
-                len(self._model_data.coords['techs_transmission_names'].values)
+                len(self._model_data.coords.get('techs_non_transmission', [])) +
+                len(self._model_data.coords.get('techs_transmission_names', []))
             ),
-            times=len(self._model_data.coords['timesteps'].values))
+            times=len(self._model_data.coords.get('timesteps', [])))
         info_strings.append('Model size:   {}'.format(msize))
         return '\n'.join(info_strings)
