### 2018-05-04 fbb3097
Plotting fixes
Allow plotting for:

* inputs-only models,
* single location models,
* models without location coordinates
