### 2018-05-04 0e53f14
Add constraints & variables to track storage between clustered datesteps (#106)
Also adds cyclic storage, which (if enabled) links the storage at the end of the whole (unclustered) timeseries to the storage in the initial step.
