### 2018-05-04 62ff593
Fix negative om_con costs in conversion techs
