### 2018-05-08 40cf5f1
Hyperlink DOIs against preferred resolver
