### 2018-05-11 ed178a7
Separate `primary_carrier` into `primary_carrier_in`/`_out` for conversion_plus techs, to allow `om_con` and `om_prod` costs
