### 2018-05-11 ce2a7e6
Add further tests & fix Hound issues
