### 2018-05-21 d0afa0e
run_scenario in cli; keep results of new scenarios; fix unsorted lists
