+++ calliope/config/model.yaml
@@ -36,6 +36,7 @@ model:
     random_seed: null  # Seed for random number generator used during clustering
     time: {}  # Optional settings to adjust time resolution, see :ref:`time_clustering` for the available options
     group_share: {}  # Optional settings for the group_share constraint
+    probability: 1  # Probability of scenarios for scenario optimisation. Sum of probability for all scenarios should equal 1
 
 ##
 # Base technology groups
