### 2018-05-21 c46c081
Update get_param in Pyomo backend
