+++ calliope/backend/pyomo/constraints/conversion_plus.py
@@ -101,13 +101,15 @@ def balance_conversion_plus_primary_constraint_rule(backend_model, loc_tech, sce
         model_data_dict['lookup_loc_techs_conversion_plus']['in', loc_tech]
     )
 
-    energy_eff = get_param(backend_model, 'energy_eff', (loc_tech, scenario, timestep))
+    energy_eff = get_param(
+        backend_model, 'energy_eff', loc_techs=loc_tech, scenarios=scenario, timesteps=timestep
+    )
 
     carrier_prod = sum(backend_model.carrier_prod[loc_tech_carrier, scenario, timestep]
-                 / get_param(backend_model, 'carrier_ratios', ('out', loc_tech_carrier))
+                 / model_data_dict['carrier_ratios']['out', loc_tech_carrier]
                  for loc_tech_carrier in loc_tech_carriers_out)
     carrier_con = sum(backend_model.carrier_con[loc_tech_carrier, scenario, timestep]
-                 * get_param(backend_model, 'carrier_ratios', ('in', loc_tech_carrier))
+                 * model_data_dict['carrier_ratios']['in', loc_tech_carrier]
                  for loc_tech_carrier in loc_tech_carriers_in)
 
     return carrier_prod == -1 * carrier_con * energy_eff
@@ -158,14 +160,17 @@ def carrier_production_min_conversion_plus_constraint_rule(backend_model, loc_te
     model_data_dict = backend_model.__calliope_model_data__['data']
 
     timestep_resolution = backend_model.timestep_resolution[timestep]
-    min_use = get_param(backend_model, 'energy_cap_min_use', (loc_tech, scenario, timestep))
+    min_use = get_param(
+        backend_model, 'energy_cap_min_use',
+        loc_techs=loc_tech, scenarios=scenario, timesteps=timestep
+    )
 
     loc_tech_carriers_out = split_comma_list(
         model_data_dict['lookup_loc_techs_conversion_plus']['out', loc_tech]
     )
 
     carrier_prod = sum(backend_model.carrier_prod[loc_tech_carrier, scenario, timestep]
-                for loc_tech_carrier in loc_tech_carriers_out)
+                       for loc_tech_carrier in loc_tech_carriers_out)
 
     return carrier_prod >= (
         timestep_resolution * backend_model.energy_cap[loc_tech] * min_use
@@ -198,8 +203,10 @@ def cost_var_conversion_plus_constraint_rule(backend_model, cost, loc_tech, scen
     var_cost = 0
 
     if loc_tech_carrier in backend_model.loc_tech_carriers_prod:
-        cost_om_prod = get_param(backend_model, 'cost_om_prod',
-                                 (cost, loc_tech, scenario, timestep))
+        cost_om_prod = get_param(
+            backend_model, 'cost_om_prod', costs=cost, loc_techs=loc_tech,
+            scenarios=scenario, timesteps=timestep
+        )
         if cost_om_prod:
             var_cost += (
                 cost_om_prod * weight *
@@ -207,8 +214,10 @@ def cost_var_conversion_plus_constraint_rule(backend_model, cost, loc_tech, scen
             )
 
     if loc_tech_carrier in backend_model.loc_tech_carriers_con:
-        cost_om_con = get_param(backend_model, 'cost_om_con',
-                                (cost, loc_tech, scenario, timestep))
+        cost_om_con = get_param(
+            backend_model, 'cost_om_con', costs=cost, loc_techs=loc_tech,
+            scenarios=scenario, timesteps=timestep
+        )
         if cost_om_con:
             var_cost += (
                 -1 * cost_om_con * weight *
[…]
