+++ calliope/backend/pyomo/constraints/milp.py
@@ -190,8 +190,11 @@ def carrier_production_max_milp_constraint_rule(backend_model, loc_tech_carrier,
 
     carrier_prod = backend_model.carrier_prod[loc_tech_carrier, scenario, timestep]
     timestep_resolution = backend_model.timestep_resolution[timestep]
-    parasitic_eff = get_param(backend_model, 'parasitic_eff', (loc_tech, scenario, timestep))
-    energy_cap = get_param(backend_model, 'energy_cap_per_unit', loc_tech)
+    parasitic_eff = get_param(
+        backend_model, 'parasitic_eff',
+        loc_techs=loc_tech, scenarios=scenario, timesteps=timestep
+    )
+    energy_cap = get_param(backend_model, 'energy_cap_per_unit', loc_techs=loc_tech)
 
     return carrier_prod <= (
         backend_model.operating_units[loc_tech, scenario, timestep] *
@@ -217,7 +220,7 @@ def carrier_production_max_conversion_plus_milp_constraint_rule(backend_model, l
     """
     model_data_dict = backend_model.__calliope_model_data__['data']
     timestep_resolution = backend_model.timestep_resolution[timestep]
-    energy_cap = get_param(backend_model, 'energy_cap_per_unit', loc_tech)
+    energy_cap = get_param(backend_model, 'energy_cap_per_unit', loc_techs=loc_tech)
     loc_tech_carriers_out = (
         split_comma_list(model_data_dict['lookup_loc_techs_conversion_plus']['out', loc_tech])
     )
@@ -248,8 +251,11 @@ def carrier_production_min_milp_constraint_rule(backend_model, loc_tech_carrier,
     loc_tech = get_loc_tech(loc_tech_carrier)
     carrier_prod = backend_model.carrier_prod[loc_tech_carrier, scenario, timestep]
     timestep_resolution = backend_model.timestep_resolution[timestep]
-    min_use = get_param(backend_model, 'energy_cap_min_use', (loc_tech, scenario, timestep))
-    energy_cap = get_param(backend_model, 'energy_cap_per_unit', loc_tech)
+    min_use = get_param(
+        backend_model, 'energy_cap_min_use',
+        loc_techs=loc_tech, scenarios=scenario, timesteps=timestep
+    )
+    energy_cap = get_param(backend_model, 'energy_cap_per_unit', loc_techs=loc_tech)
 
     return carrier_prod >= (
         backend_model.operating_units[loc_tech, scenario, timestep] *
@@ -275,8 +281,11 @@ def carrier_production_min_conversion_plus_milp_constraint_rule(backend_model, l
     """
     model_data_dict = backend_model.__calliope_model_data__['data']
     timestep_resolution = backend_model.timestep_resolution[timestep]
-    energy_cap = get_param(backend_model, 'energy_cap_per_unit', loc_tech)
-    min_use = get_param(backend_model, 'energy_cap_min_use', (loc_tech, scenario, timestep))
+    energy_cap = get_param(backend_model, 'energy_cap_per_unit', loc_techs=loc_tech)
+    min_use = get_param(
+        backend_model, 'energy_cap_min_use',
+        loc_techs=loc_tech, scenarios=scenario, timesteps=timestep
+    )
     loc_tech_carriers_out = (
         split_comma_list(model_data_dict['lookup_loc_techs_conversion_plus']['out', loc_tech])
     )
@@ -308,7 +317,7 @@ def carrier_consumption_max_milp_constraint_rule(backend_model, loc_tech_carrier
     loc_tech = get_loc_tech(loc_tech_carrier)
     carrier_con = backend_model.carrier_con[loc_tech_carrier, scenario, timestep]
[…]
