### 2018-05-22 7e98044
`robust_plan` -> `scenario_plan` mode name
