+++ calliope/backend/pyomo/variables.py
@@ -98,7 +98,7 @@ def initialize_decision_variables(backend_model):
 def initialize_cvar_decision_variables(backend_model):
     """
     Define decision variables specific to managing conditional value at risk
-    (function is triggered if user selects 'robust_plan' as the model mode)
+    (function is triggered if user selects 'scenario_plan' as the model mode)
     """
     # Load auxiliary decision variables used to calculate the CVaR
     backend_model.xi = po.Var(within=po.Reals)
