### 2018-05-22 9deaae1
Update error raised by multiple scenarios in plan mode
