### 2018-05-29 e895d6d
Ensure floating point error doesn't break sum(probability) == 1
