### 2018-05-30 0455ec4
Allow YAML comments to pass through AttrDict
