+++ calliope/core/attrdict.py
@@ -10,24 +10,37 @@ used for managing model configuration.
 
 """
 
-import copy
-import logging
-
 import numpy as np
-import ruamel.yaml as yaml
+import ruamel.yaml as ruamel_yaml
 
 from calliope.core.util.tools import relative_path
 from calliope.core.util.logging import logger
 
-class __Missing(object):
-    def __repr__(self):
-        return ('MISSING')
 
+class __Missing(object):
     def __nonzero__(self):
         return False
 
 
 _MISSING = __Missing()
+
+
+class CalliopeYAMLEmitter(ruamel_yaml.emitter.Emitter):
+
+    def write_comment(self, comment):
+        """
+        Override write_comment such as to ensure that inline
+        comments are two spaces from the end of a statement,
+        while full-line comments are aligned with their line.
+
+        """
+        comment_too_close = comment.start_mark.column < self.column
+        comment_too_far = comment.start_mark.column > self.column + 2
+
+        if comment_too_close or comment_too_far:
+            comment.start_mark.column = self.column + 2
+
+        super().write_comment(comment)
 
 
 def _yaml_load(src):
@@ -43,9 +56,17 @@ def _yaml_load(src):
     else:
         src_name = '<yaml string>'
     try:
-        return yaml.safe_load(src)
-    except yaml.YAMLError:
-        logger.error('Parser error when reading YAML from {}.'.format(src_name))
+        # Create a ruamel_yaml.YAML instance in order to get the
+        # comment roundtripping that it provides by default
+        yaml_ = ruamel_yaml.YAML()
+        result = yaml_.load(src)
+        if not isinstance(result, dict):
+            raise ValueError('Could not parse {} as YAML'.format(src_name))
+        return result
+    except ruamel_yaml.YAMLError:
+        logger.error(
+            'Parser error when reading YAML '
+            'from {}.'.format(src_name))
         raise
 
 
@@ -62,19 +83,63 @@ class AttrDict(dict):
     """
     __getattr__ = dict.__getitem__
     __setattr__ = dict.__setitem__
-    __delattr__ = dict.__delitem__
+
+    # We add a dict key that contains comment metadata, but override
+    # the __iter__, keys, and items methods below in order to hide
+    # this key from users -- it remains accessible with direct key
+    # access but is not listed or iterated over
+    __hidden_keys__ = ['__dict_comments__']
 
     def __init__(self, source_dict=None):
-        super(AttrDict, self).__init__()
-        if isinstance(source_dict, dict):
-            self.init_from_dict(source_dict)
+        super().__init__()
 
-    def copy(self, deep=False):
-        """Override copy method so that it returns an AttrDict"""
-        if deep:
-            return AttrDict(copy.deepcopy(self.as_dict()))
+        setattr(self, '__dict_comments__', {})
+
+        if source_dict is not None:
+            if isinstance(source_dict, dict):
+                self.init_from_dict(source_dict)
+            else:
+                raise ValueError('Must pass a dict to AttrDict')
+
+    def __iter__(self):
+        return iter(self.keys())
+
+    def __delitem__(self, key):
+        try:
+            del self['__dict_comments__'][key]
+        except KeyError:
+            pass
+        super().__delitem__(key)
+
+    __delattr__ = __delitem__
+
+    def keys(self, filtered=True):
+        if filtered:
+            return [
+                i for i in super().keys()
+                if i not in self.__hidden_keys__]
         else:
-            return AttrDict(self.as_dict().copy())
+            return super().keys()
+
+    def values(self, filtered=True):
+        if filtered:
+            return [
+                i[1] for i in super().items()
+                if i[0] not in self.__hidden_keys__]
+        else:
+            return super().values()
+
+    def items(self, filtered=True):
+        if filtered:
+            return [
+                i for i in super().items()
[…]
