+++ calliope/test/test_core_attrdict.py
@@ -1,8 +1,13 @@
 from io import StringIO
 import os
-import pytest
 
-from calliope.core.attrdict import AttrDict
+import pytest
+import numpy as np
+import tempfile
+import ruamel.yaml as ruamel_yaml
+
+from calliope.core.attrdict import AttrDict, _MISSING
+from calliope.test.common.util import check_error_or_warning
 
 
 class TestAttrDict:
@@ -20,11 +25,15 @@ class TestAttrDict:
         return d
 
     setup_string = """
+        # a comment
         a: 1
         b: 2
-        c:
-            x: foo
-            y: bar
+        # a comment about `c`
+        c:  # a comment inline with `c`
+            x: foo  # a comment on foo
+
+            #
+            y: bar  #
             z:
                 I: 1
                 II: 2
@@ -43,6 +52,19 @@ class TestAttrDict:
     def attr_dict(self, regular_dict):
         d = regular_dict
         return AttrDict(d)
+
+    def test_missing_nonzero(self):
+        assert _MISSING is not True
+        assert _MISSING is not False
+        assert _MISSING is not None
+        assert _MISSING.__nonzero__() is False
+
+    def test_init_from_nondict(self):
+        with pytest.raises(ValueError) as excinfo:
+            d = AttrDict('foo')
+        assert check_error_or_warning(
+            excinfo, 'Must pass a dict to AttrDict'
+        )
 
     def test_init_from_dict(self, regular_dict):
         d = AttrDict(regular_dict)
@@ -79,6 +101,24 @@ class TestAttrDict:
         yaml_string = 'a.b.c: 1\na.b.c: 2'
         d = AttrDict.from_yaml_string(yaml_string)
         assert d.a.b.c == 2
+
+    def test_simple_invalid_yaml(self):
+        yaml_string = '1 this is not valid yaml'
+        with pytest.raises(ValueError) as excinfo:
+            d = AttrDict.from_yaml_string(yaml_string)
+        assert check_error_or_warning(
+            excinfo, 'Could not parse <yaml string> as YAML'
+        )
+
+    def test_parser_error(self):
+        with pytest.raises(ruamel_yaml.YAMLError):
+            d = AttrDict.from_yaml_string("""
+            foo: bar
+            baz: 1
+                - foobar
+                bar: baz
+
+            """)
 
     def test_dot_access_first(self, attr_dict):
         d = attr_dict
@@ -118,6 +158,25 @@ class TestAttrDict:
         d.set_key('d.foo', 'bar')
         assert d.d.foo == 'bar'
 
+    def test_pass_regular_dict_to_set_key(self, attr_dict):
+        # Regular dicts get turned into AttrDicts when using
+        # assignment through set_key()
+        attr_dict.set_key('c.z.newkey', {'foo': 1, 'doo': 2})
+        assert isinstance(attr_dict.get_key('c.z.newkey'), AttrDict)
+        assert attr_dict.get_key('c.z.newkey.foo') == 1
+
+    def test_get_subkey_from_nested_non_attrdict(self, attr_dict):
+        # Directly assigning a dict means it is not modified
+        # but it breaks get_key with nested keys
+        attr_dict['c']['z']['newkey'] = {'foo': 1, 'doo': 2}
+
+        with pytest.raises(AttributeError) as excinfo:
+            attr_dict.get_key('c.z.newkey.foo')
+
+        assert check_error_or_warning(
+            excinfo, "'dict' object has no attribute 'get_key'"
+        )
+
     def test_get_key_first(self, attr_dict):
         d = attr_dict
         assert d.get_key('a') == 1
@@ -147,6 +206,10 @@ class TestAttrDict:
     def test_get_key_second_inexistant_default(self, attr_dict):
[…]
