### 2018-05-30 0bbea92
CLI: use ipython debugger instead of built-in pdb
