### 2018-05-30 534d67a
Update changelog; fix a trailing bracket
