+++ changelog.rst
@@ -10,6 +10,8 @@ Release History
 
 |new| On clustering timeseries into representative days, an additional set of decision variables and constraints is generated. This addition allows for tracking stored energy between clusters, by considering storage between every `datestep` of the original (unclustered) timeseries as well as storage variation within a cluster.
 
+|new| AttrDict now persists comments when reading from and writing to YAML files, and gains an API to view, add and remove comments on keys
+
 |fixed| Fix CLI error when running a model without transmission technologies
 
 |fixed| Allow plotting for inputs-only models, single location models, and models without location coordinates
