### 2018-05-30 8a3df6d
Update changelog; fix a trailing bracket
