### 2018-05-30 2cc8498
Separate `primary_carrier` into `primary_carrier_in`/`_out` (#111)
Separate `primary_carrier` into `primary_carrier_in`/`_out` for conversion_plus techs, to allow `om_con` and `om_prod` costs
