### 2018-05-30 c578c19
Fix undefined variable
