### 2018-05-30 52f5444
Cosmetic fixes
* Docs: add GitHub ribbon to index page
* CLI: use ipython debugger instead of built-in pdb
