### 2018-05-30 193e6be
Remove need for scenario dimension on all timestep vars; update objective
