+++ calliope/core/preprocess/constraint_sets.py
@@ -215,10 +215,16 @@ def generate_constraint_sets(model_run):
         i for i in sets.loc_techs_milp
         if i in sets.loc_techs_investment_cost and
         any(constraint_exists(model_run, i, 'costs.{}.purchase'.format(j))
-               for j in model_run.sets.costs)
+            for j in model_run.sets.costs)
     ]
     # loc_techs_purchase technologies only exist because they have defined a purchase cost
     constraint_sets['loc_techs_update_costs_investment_purchase_constraint'] = sets.loc_techs_purchase
+
+    constraint_sets['techs_unit_capacity_systemwide_constraint'] = [
+        i for i in sets.techs
+        if model_run.get_key('techs.{}.constraints.units_max_systemwide'.format(i), None)
+        or model_run.get_key('techs.{}.constraints.units_equals_systemwide'.format(i), None)
+    ]
 
     # conversion.py
     constraint_sets['loc_techs_balance_conversion_constraint'] = sets.loc_techs_conversion
