+++ calliope/test/test_backend_pyomo.py
@@ -1232,8 +1232,49 @@ class TestMILPConstraints:
         assert check_variable_exists(m._backend_model, 'cost_investment_constraint', 'purchased')
         assert not check_variable_exists(m._backend_model, 'cost_investment_constraint', 'units')
 
+    def test_techs_unit_capacity_systemwide_constraint(self):
+        """
+        sets.techs if unit_cap_max_systemwide or unit_cap_equals_systemwide
+        """
+
+        override_max = {
+            'links.0,1.exists': True,
+            'techs.test_conversion_plus.constraints.units_max_systemwide': 2,
+            'locations.1.techs.test_conversion_plus.constraints': {
+                'units_max': 2,
+                'energy_cap_per_unit': 5
+            }
+        }
+        override_equals = {
+            'links.0,1.exists': True,
+            'techs.test_conversion_plus.constraints.units_equals_systemwide': 1,
+            'locations.1.techs.test_conversion_plus.costs.monetary.purchase': 1
+        }
+        override_equals_inf = {
+            'links.0,1.exists': True,
+            'techs.test_conversion_plus.constraints.units_equals_systemwide': np.inf,
+            'locations.1.techs.test_conversion_plus.costs.monetary.purchase': 1
+        }
+
+        m = build_model(override_max, 'conversion_plus_milp,two_hours,investment_costs')
+        m.run(build_only=True)
+        assert hasattr(m._backend_model, 'unit_capacity_systemwide_constraint')
+        assert m._backend_model.unit_capacity_systemwide_constraint['test_conversion_plus'].upper() == 2
+
+        m = build_model(override_equals, 'conversion_plus_milp,two_hours,investment_costs')
+        m.run(build_only=True)
+        assert hasattr(m._backend_model, 'unit_capacity_systemwide_constraint')
+        assert m._backend_model.unit_capacity_systemwide_constraint['test_conversion_plus'].lower() == 1
+        assert m._backend_model.unit_capacity_systemwide_constraint['test_conversion_plus'].upper() == 1
+
+        with pytest.raises(ValueError) as error:
+            m = build_model(override_equals_inf, 'conversion_plus_milp,two_hours,investment_costs')
+            m.run(build_only=True)
+        assert check_error_or_warning(error, 'Cannot use inf for energy_cap_equals_systemwide')
+
 
 class TestConversionConstraints:
+
     # conversion.py
     def test_loc_techs_balance_conversion_constraint(self):
         """
