### 2018-05-31 d7e92f6
Add docs and FutureWarning for cyclic storage; minor cleanup
