### 2018-06-01 b8592cf
Minor doc updates
