### 2018-06-01 f1c68ab
Fix carrier_prod sum check for urban-scale example
