+++ calliope/test/test_example_models.py
@@ -266,7 +266,7 @@ class TestUrbanScaleExampleModelSenseChecks:
         assert model.results.energy_cap.to_pandas()['X1::chp'] == approx(272.227204)
         assert model.results.energy_cap.to_pandas()['X2::heat_pipes:N1'] == approx(197.908691)
 
-        assert model.results.carrier_prod.sum('timesteps').to_pandas()['X3::boiler::heat'] == approx(462.189531)
+        assert model.results.carrier_prod.sum('timesteps').to_pandas()['X3::boiler::heat'] == approx(474.720800)
         assert float(model.results.carrier_export.sum()) == approx(0)
 
         assert float(model.results.cost.sum()) == approx(528.8643154)
