### 2018-06-01 53b2bc0
Fix negative om_con costs in conversion techs (#109)
