+++ calliope/config/model.yaml
@@ -44,7 +44,7 @@ model:
 tech_groups:
     supply:
         required_constraints: [['energy_cap_max', 'energy_cap_equals', 'energy_cap_per_unit']]
-        allowed_constraints: ['energy_prod', 'lifetime', 'resource', 'force_resource', 'resource_min_use', 'resource_unit', 'resource_area_min', 'resource_area_max', 'resource_area_equals', 'resource_area_per_energy_cap', 'resource_scale', 'energy_eff', 'energy_cap_min', 'energy_cap_max', 'energy_cap_equals', 'energy_cap_max_systemwide', 'energy_cap_equals_systemwide', 'energy_cap_scale', 'energy_cap_min_use', 'energy_cap_per_unit', 'energy_ramping', 'energy_eff_per_distance', 'export_cap', 'export_carrier', 'units_min', 'units_max', 'units_equals']
+        allowed_constraints: ['energy_prod', 'lifetime', 'resource', 'force_resource', 'resource_min_use', 'resource_unit', 'resource_area_min', 'resource_area_max', 'resource_area_equals', 'resource_area_per_energy_cap', 'resource_scale', 'energy_eff', 'energy_cap_min', 'energy_cap_max', 'energy_cap_equals', 'energy_cap_max_systemwide', 'energy_cap_equals_systemwide', 'energy_cap_scale', 'energy_cap_min_use', 'energy_cap_per_unit', 'energy_ramping', 'energy_eff_per_distance', 'export_cap', 'export_carrier', 'units_min', 'units_max', 'units_equals', 'units_max_systemwide', 'units_equals_systemwide']
         allowed_costs: ['interest_rate', 'resource_area', 'energy_cap', 'om_annual_investment_fraction', 'om_annual', 'om_prod', 'om_con', 'export', 'purchase', 'depreciation_rate']
         essentials:
             parent: null
@@ -54,7 +54,7 @@ tech_groups:
             energy_prod: true
     supply_plus:
         required_constraints: [['energy_cap_max', 'energy_cap_equals', 'energy_cap_per_unit']]
-        allowed_constraints: ['energy_prod', 'lifetime', 'resource', 'force_resource', 'resource_min_use', 'resource_unit', 'resource_eff', 'resource_area_min', 'resource_area_max', 'resource_area_equals', 'resource_area_per_energy_cap', 'resource_cap_min', 'resource_cap_max', 'resource_cap_equals', 'resource_cap_equals_energy_cap', 'resource_scale', 'parasitic_eff', 'energy_eff', 'energy_cap_min', 'energy_cap_max', 'energy_cap_equals', 'energy_cap_max_systemwide', 'energy_cap_equals_systemwide', 'energy_cap_scale', 'energy_cap_min_use', 'energy_cap_per_unit', 'energy_ramping', 'energy_eff_per_distance', 'export_cap', 'export_carrier', 'units_min', 'units_max', 'units_equals', 'storage_initial', 'storage_cap_min', 'storage_cap_max', 'storage_cap_equals', 'storage_cap_per_unit', 'charge_rate', 'storage_time_max', 'storage_loss']
[…]
