### 2018-06-01 e96b247
Update where we store & check for FutureWarnings and DeprecationWarnings
