+++ calliope/core/preprocess/checks.py
@@ -555,3 +555,4 @@ def check_future_deprecation_warnings(model_run, model_data):
             'From v0.6.3, cyclic storage will default to True.',
             FutureWarning
         )
+
