### 2018-06-04 af2a4de
remove pointless line (silly Hound)
