### 2018-06-04 8cbbe73
Update where we store & check for FutureWarnings and DeprecationWarnings (#117)
Fixes issue #108
