+++ calliope/core/preprocess/checks.py
@@ -10,7 +10,7 @@ Checks for model consistency and possible errors during preprocessing.
 """
 
 import os
-import logging
+import warnings
 
 import numpy as np
 import xarray as xr
@@ -37,7 +37,7 @@ def check_overrides(config_model, override):
     Perform checks on the override dict and override file inputs to ensure they
     are not doing something silly.
     """
-    warnings = []
+    model_warnings = []
     info = []
     for key in override.as_dict_flat().keys():
         if key in config_model.as_dict_flat().keys():
@@ -70,7 +70,7 @@ def check_overrides(config_model, override):
         if config_coordinates != override_coordinates:
             for key in config_keys:
                 config_model.del_key(key)
-            warnings.append(
+            model_warnings.append(
                 'Updated from coordinate system {} to {}, using overrides'
                 .format(config_coordinates, override_coordinates)
             )
@@ -78,7 +78,7 @@ def check_overrides(config_model, override):
     if info:
         logger.info('\n'.join(info))
 
-    return warnings
+    return model_warnings
 
 
 def check_initial(config_model):
@@ -87,7 +87,7 @@ def check_initial(config_model):
 
     Returns
     -------
-    warnings : list
+    model_warnings : list
         possible problems that do not prevent the model run
         from continuing
     errors : list
@@ -95,13 +95,13 @@ def check_initial(config_model):
 
     """
     errors = []
-    warnings = []
+    model_warnings = []
 
     # Check for version mismatch
     model_version = config_model.model.get('calliope_version', False)
     if model_version:
         if not str(model_version) in __version__:
-            warnings.append(
+            model_warnings.append(
                 'Model configuration specifies calliope_version={}, '
                 'but you are running {}. Proceed with caution!'.format(
                     model_version, __version__)
@@ -110,7 +110,7 @@ def check_initial(config_model):
     # Check top-level keys
     for k in config_model.keys():
         if k not in ['model', 'run', 'locations', 'tech_groups', 'techs', 'links', 'config_path']:
-            warnings.append(
+            model_warnings.append(
                 'Unrecognised top-level configuration item: {}'.format(k)
             )
 
@@ -120,14 +120,14 @@ def check_initial(config_model):
     for k in config_model['run'].keys_nested():
         if (k not in defaults_model['run'].keys_nested() and
                 'solver_options' not in k):
-            warnings.append(
+            model_warnings.append(
                 'Unrecognised setting in run configuration: {}'.format(k)
             )
 
     # Check model configuration, but top-level keys only
     for k in config_model['model'].keys():
         if k not in defaults_model['model'].keys():
-            warnings.append(
+            model_warnings.append(
                 'Unrecognised setting in model configuration: {}'.format(k)
             )
 
@@ -220,7 +220,7 @@ def check_initial(config_model):
             )
     for arg in config_model.run.objective_options:
         if arg not in objective_args_expected:
-            warnings.append(
+            model_warnings.append(
                 'Objective function argument `{}` given but not used by objective function `{}`'
                 .format(arg, config_model.run.objective)
             )
@@ -236,20 +236,20 @@ def check_initial(config_model):
             '`storage_inter_cluster` decision variable is not activated.'
         )
 
-    return warnings, errors
+    return model_warnings, errors
 
 
-def _check_tech(model_run, tech_id, tech_config, loc_id, warnings, errors, comments):
+def _check_tech(model_run, tech_id, tech_config, loc_id, model_warnings, errors, comments):
     """
     Checks individual tech/tech groups at specific locations.
-    NOTE: Updates `warnings` and `errors` lists in-place.
[…]
