+++ calliope/core/model.py
@@ -20,6 +20,7 @@ from calliope.core.preprocess import \
     build_model_data, \
     apply_time_clustering, \
     final_timedimension_processing
+from calliope.core.preprocess.checks import check_future_deprecation_warnings
 from calliope.core.util.logging import log_time
 from calliope.core.util.dataset import split_loc_techs
 from calliope.core.util.tools import apply_to_dict
@@ -77,6 +78,7 @@ class Model(object):
             raise ValueError(
                 'Input configuration must either be a string or a dictionary.'
             )
+        check_future_deprecation_warnings(self._model_run, self._model_data)
 
         self.plot = plotting.ModelPlotMethods(self)
 
