+++ calliope/core/preprocess/time.py
@@ -238,15 +238,4 @@ def final_timedimension_processing(model_data):
     model_data = reorganise_dataset_dimensions(model_data)
     model_data = add_max_demand_timesteps(model_data)
 
-
-    ## Warning that cyclic storage will default to True in 0.6.3 ####
-    # TODO: remove in v0.6.3-dev
-    if 'loc_techs_store' in model_data and not model_data.attrs.get('run.cyclic_storage', False):
-        warnings.warn(
-            'Cyclic storage, a new addition in v0.6.2, currently defaults to '
-            'False (i.e. emulating functionality prior to v0.6.2). '
-            'From v0.6.3, cyclic storage will default to True.',
-            FutureWarning
-        )
-
     return model_data
