### 2018-06-04 4c6e4ea
Update doc plots for new release; minor changelog updates
