### 2018-06-11 67a125e
Fix bug when running without scenarios from cli
