### 2018-06-11 c1d7cc2
Add SLURM (sbatch) script generator for use on a remote cluster
