+++ changelog.rst
@@ -3,6 +3,9 @@
 Release History
 ===============
 
+0.6.3-dev
+---------
+
 0.6.2 (2018-06-04)
 ------------------
 
