### 2018-06-11 462cd9b
Remove Cambridge-specific SLURM script
