### 2018-06-11 d0741b1
Update output log file naming from SLURM runs
