### 2018-06-11 af939f4
Add SLURM to script generator (#118)
