+++ calliope/core/util/generate_runs.py
@@ -12,6 +12,7 @@ in parallel on a cluster or sequentially on any machine.
 
 import os
 
+import pandas as pd
 from calliope.core import AttrDict
 
 
@@ -71,15 +72,17 @@ def generate_bash_script(out_file, model_file, override_file, groups, additional
 
     lines_all = lines_start + [base_string.format(i=i + 1, cmd=cmd) for i, cmd in enumerate(commands)] + lines_end
 
-    with open(out_file, 'w') as f:
-        f.write('\n'.join(lines_all))
+    with open(out_file, 'wb') as f:
+        f.write(bytes('\n'.join(lines_all), 'UTF-8'))
 
     os.chmod(out_file, 0o755)
 
     return commands
 
 
-def generate_bsub_script(out_file, model_file, override_file, groups, additional_args, cluster_mem, cluster_time, cluster_threads=1, **kwargs):
+def generate_bsub_script(out_file, model_file, override_file, groups,
+                         additional_args, cluster_mem, cluster_time,
+                         cluster_threads=1, **kwargs):
 
     # We also need to generate the bash script to run on the cluster
     bash_out_file = out_file + '.array.sh'
@@ -99,8 +102,55 @@ def generate_bsub_script(out_file, model_file, override_file, groups, additional
         ''
     ]
 
-    with open(out_file, 'w') as f:
-        f.write('\n'.join(lines))
+    with open(out_file, 'wb') as f:
+        f.write(bytes('\n'.join(lines), 'UTF-8'))
+
+
+def generate_sbatch_script(out_file, model_file, override_file, groups,
+                           additional_args, cluster_mem, cluster_time,
+                           cluster_threads=1, **kwargs):
+    """
+    SBATCH (SLURM) script generator.
+    """
+
+    # We also need to generate the bash script to run on the cluster
+    bash_out_file = out_file + '.array.sh'
+    bash_out_file_basename = os.path.basename(bash_out_file)
+    commands = generate_bash_script(bash_out_file, model_file, override_file, groups, additional_args)
+
+    if ':' not in cluster_time:
+        # Assuming time given as minutes, so needs changing to %H:%M%S
+        cluster_time = pd.to_datetime(cluster_time, unit='m').strftime('%H:%M:%S')
+
+    lines = [
+        '#!/bin/bash',
+        '#SBATCH -J calliope',  # Name of the job
+        '#SBATCH --array=1-{}'.format(len(commands)),  # How many jobs there are
+        # How many (MPI) tasks will there be in total? (<= nodes*32)
+        # The skylake/skylake-himem nodes have 32 CPUs (cores) each.
+        '#SBATCH --ntasks={}'.format(cluster_threads),
+        '#SBATCH --time={}'.format(cluster_time),  # How much wallclock time will be required
+        '#SBATCH -o log_%a.log',
+        '',
+        '#! Optional add-ins for SBATCH (uncomment and add info as necessary):',
+        '##SBATCH -A project_name',  # Which project should be charged'
+        '##SBATCH --nodes=X',  # X whole nodes should be allocated'
+        '##SBATCH -p partition_name',
+        '',
+        '#! Insert module load commands after this line, if needed:',
+        '#! (Note: you can load this in ~.bashrc if you want them loaded every time you log in)',
+        '',
+        '#! module load gurobi (or glpk/cplex)',
+        '#! module load miniconda3',
+        '#! module load /path/to/miniconda3/envs/your_env_name/',
+        '',
+        'cd $SLURM_SUBMIT_DIR',
+        '',
[…]
