### 2018-06-11 f187f1f
Add `--mem` option to SLURM run generation script
