### 2018-06-12 12132db
Fix missing `scenario` dimension in `carrier_con` index
