+++ calliope/backend/pyomo/constraints/conversion_plus.py
@@ -232,7 +232,7 @@ def cost_var_conversion_plus_constraint_rule(backend_model, cost, loc_tech, scen
         if cost_om_con:
             var_cost += (
                 cost_om_con * weight * -1 *
-                backend_model.carrier_con[loc_tech_carrier_con, timestep]
+                backend_model.carrier_con[loc_tech_carrier_con, scenario, timestep]
             )
 
     backend_model.cost_var_rhs[cost, loc_tech, scenario, timestep] = var_cost
