### 2018-06-13 163cce7
Add transmission tech test for `unit_cap_systemwide`
