+++ calliope/backend/pyomo/constraints/milp.py
@@ -566,10 +566,18 @@ def unit_capacity_systemwide_constraint_rule(backend_model, tech):
 
     """
 
-    all_loc_techs = [
-        i for i in backend_model.loc_techs
-        if i.split('::')[1] == tech
-    ]
+    if tech in backend_model.techs_transmission_names:
+        all_loc_techs = [
+            i for i in backend_model.loc_techs_transmission
+            if i.split('::')[1].split(':')[0] == tech
+        ]
+        multiplier = 2  # there are always two technologies associated with one link
+    else:
+        all_loc_techs = [
+            i for i in backend_model.loc_techs
+            if i.split('::')[1] == tech
+        ]
+        multiplier = 1
 
     max_systemwide = get_param(backend_model, 'units_max_systemwide', tech)
     equals_systemwide = get_param(backend_model, 'units_equals_systemwide', tech)
@@ -590,9 +598,7 @@ def unit_capacity_systemwide_constraint_rule(backend_model, tech):
         if loc_tech_is_in(backend_model, loc_tech, 'loc_techs_purchase')
     )
 
-    total_expr = equals_systemwide if equals_systemwide else max_systemwide
-
     if equals_systemwide:
-        return sum_expr_units + sum_expr_purchase == total_expr
+        return sum_expr_units + sum_expr_purchase == equals_systemwide * multiplier
     else:
-        return sum_expr_units + sum_expr_purchase <= total_expr
+        return sum_expr_units + sum_expr_purchase <= max_systemwide * multiplier
