### 2018-06-13 93e9d1a
required_constraints must also be allowed_constraints (#120)
