### 2018-06-13 05564ca
Alphabetically sort `model` and `defaults` config YAML files
