+++ calliope/config/defaults.yaml
@@ -20,79 +20,79 @@
 #                 one_way: false  # FIXME needs description
 
 # List of constraints allowed to specify "file="
-file_allowed: ['resource', 'resource_eff', 'parasitic_eff', 'energy_eff', 'storage_loss', 'energy_ramping', 'energy_prod', 'energy_con', 'force_resource', 'om_prod', 'om_con', 'export', 'clustering_func']
+file_allowed: ['clustering_func', 'energy_con', 'energy_eff', 'energy_prod', 'energy_ramping', 'export', 'force_resource', 'om_con', 'om_prod', 'parasitic_eff', 'resource_eff', 'resource', 'storage_loss']
 
 default_location:
     available_area: null
 default_tech:
     essentials:
-        parent: null  # Parent tech_group. Must always be defined
-        name: 'Default technology'  # Name of tech, used for plotting and analysis
-        color: false  # Color to use when plotting results. If not defined, a random one will be chosen
-        primary_carrier_out: false  # Selects the primary output carrier to associate with costs & constraints (if multiple primary output carriers are assigned)
-        primary_carrier_in: false  # Selects the primary input carrier to associate with costs & constraints (if multiple primary input carriers are assigned)
         carrier_in: false  # Defaults to false, allows technologies to define primary carrier(s) to consume
         carrier_in_2: false  # Defaults to false, allows technologies to define secondary carrier(s) to consume
         carrier_in_3: false  # Defaults to false, allows technologies to define tertiary carrier(s) to consume
         carrier_out: false  # Defaults to false, allows conversion technologies to define primary carrier(s) to produce
         carrier_out_2: false  # Defaults to false, allows conversion technologies to define secondary carrier(s) to produce
         carrier_out_3: false  # Defaults to false, allows conversion technologies to define tertiary carrier(s) to produce
+        color: false  # Color to use when plotting results. If not defined, a random one will be chosen
+        name: 'Default technology'  # Name of tech, used for plotting and analysis
+        parent: null  # Parent tech_group. Must always be defined
+        primary_carrier_in: false  # Selects the primary input carrier to associate with costs & constraints (if multiple primary input carriers are assigned)
+        primary_carrier_out: false  # Selects the primary output carrier to associate with costs & constraints (if multiple primary output carriers are assigned)
     constraints:
-        lifetime: null  # name: Technology lifetime ¦ unit: years ¦ Must be defined if fixed capital costs are defined. A reasonable value for many technologies is around 20-25 years.
         carrier_ratios: {}  # name: Carrier ratios ¦ unit: fraction ¦ Ratio of summed output of carriers in ['out_2', 'out_3'] / ['in_2', 'in_3'] to the summed output of carriers in 'out' / 'in'. given in a nested dictionary.
-        resource: 0  # name: Available resource ¦ unit: kWh/m\ :sup:`2` | kW/m\ :sup:`2` ¦ Maximum available resource (static, or from file as timeseries). Unit dictated by ``reosurce_unit``
-        force_resource: false  # name: Force resource ¦ unit: boolean ¦ Forces this technology to use all available ``resource``, rather than making it a maximum upper boundary (for production) or minimum lower boundary (for consumption). Static boolean, or from file as timeseries
[…]
