+++ calliope/config/model.yaml
@@ -8,64 +8,62 @@
 # Model-wide default settings
 ##
 run:
+
     backend: pyomo  # Backend to use to build and solve the model. As of v0.6.0, only `pyomo` is available
-    solver: glpk  # Which solver to use
-    solver_options: null  # A list of options, which are passed on to the chosen solver, and are therefore solver-dependent
-    solver_io: null  # What method the Pyomo backend should use to communicate with the solver
-    save_logs: null  # Directory into which to save logs and temporary files. Also turns on symbolic solver labels in the Pyomo backend
     bigM: 1e9 # Used for unmet demand, but should be of a similar order of magnitude as the largest cost that the model could achieve. Too high and the model will not converge
+    cyclic_storage: false # If true, storage in the last timestep of the timeseries is considered to be the 'previous timestep' in the first timestep of the timeseries
     ensure_feasibility: false # If true, unmet_demand will be a decision variable, to account for an ability to meet demand with the available supply. If False and a mismatch occurs, the optimisation will fail due to infeasibility
+    mode: plan  # Which mode to run the model in: 'plan' or 'operation'
+    objective_options: {'cost_class': 'monetary', 'sense': 'minimize'}  # Arguments to pass to objective function. If cost-based objective function in use, should include 'cost_class' and 'sense' (maximize/minimize)
+    objective:  minmax_cost_optimization # Name of internal objective function to use, currently only min/max cost-based optimisation is available
     operation:  # Settings for operational mode
         window: null
         horizon: null
         use_cap_results: false
-    objective:  minmax_cost_optimization # Name of internal objective function to use, currently only min/max cost-based optimisation is available
-    objective_options: {'cost_class': 'monetary', 'sense': 'minimize'}  # Arguments to pass to objective function. If cost-based objective function in use, should include 'cost_class' and 'sense' (maximize/minimize)
-    mode: plan  # Which mode to run the model in: 'plan' or 'operation'
+    save_logs: null  # Directory into which to save logs and temporary files. Also turns on symbolic solver labels in the Pyomo backend
+    solver_io: null  # What method the Pyomo backend should use to communicate with the solver
+    solver_options: null  # A list of options, which are passed on to the chosen solver, and are therefore solver-dependent
+    solver: glpk  # Which solver to use
     zero_threshold: 1e-10 # Any value coming out of the backend that is smaller than this threshold (due to floating point errors, probably) will be set to zero
-    cyclic_storage: false # If true, storage in the last timestep of the timeseries is considered to be the 'previous timestep' in the first timestep of the timeseries
 
 model:
-    name: null  # Model name
     calliope_version: null  # Calliope framework version this model is intended for
-    timeseries_data_path: null  # Path to time series data
-    timeseries_dateformat: '%Y-%m-%d %H:%M:%S'  # Timestamp format of all time series data when read from file
-    timeseries_data: null  # Dict of dataframes with time series data (when passing in dicts rather than YAML files to Model constructor)
[…]
