### 2018-06-13 21d3970
Fix alphabetic inconsistencies
