+++ calliope/config/model.yaml
@@ -63,14 +63,14 @@ tech_groups:
             energy_prod: true
     demand:
         required_constraints: ['resource']
-        allowed_constraints: ['energy_con', 'resource', 'force_resource', 'resource_area_equals', 'resource_scale', 'resource_unit']
+        allowed_constraints: ['energy_con', 'force_resource', 'resource', 'resource_area_equals', 'resource_scale', 'resource_unit']
         allowed_costs: []
         essentials:
             parent: null
         constraints:
-            resource_unit: power
-            force_resource: true
             energy_con: true
+            force_resource: true
+            resource_unit: power
     storage:
         required_constraints: [['energy_cap_equals', 'energy_cap_max', 'energy_cap_per_unit'], ['storage_cap_equals', 'storage_cap_max']]
         allowed_constraints: ['charge_rate', 'energy_cap_equals', 'energy_cap_equals_systemwide', 'energy_cap_max', 'energy_cap_max_systemwide', 'energy_cap_min', 'energy_cap_min_use', 'energy_cap_per_unit', 'energy_cap_scale', 'energy_con', 'energy_eff', 'energy_prod', 'energy_ramping', 'export_cap', 'export_carrier', 'lifetime', 'storage_cap_equals', 'storage_cap_max', 'storage_cap_min', 'storage_cap_per_unit', 'storage_initial', 'storage_loss', 'storage_time_max', 'units_equals', 'units_equals_systemwide', 'units_max', 'units_max_systemwide', 'units_min']
@@ -89,8 +89,8 @@ tech_groups:
             parent: null
         constraints:
             energy_prod: true
-            resource_unit: power
             resource: inf
+            resource_unit: power
     supply_plus:
         required_constraints: [['energy_cap_equals', 'energy_cap_max', 'energy_cap_per_unit']]
         allowed_constraints: ['charge_rate', 'energy_cap_equals', 'energy_cap_equals_systemwide', 'energy_cap_max', 'energy_cap_max_systemwide', 'energy_cap_min', 'energy_cap_min_use', 'energy_cap_per_unit', 'energy_cap_scale', 'energy_eff', 'energy_prod', 'energy_ramping', 'export_cap', 'export_carrier', 'force_resource', 'lifetime', 'parasitic_eff', 'resource', 'resource_area_equals', 'resource_area_max', 'resource_area_min', 'resource_area_per_energy_cap', 'resource_cap_equals', 'resource_cap_equals_energy_cap', 'resource_cap_max', 'resource_cap_min', 'resource_eff', 'resource_min_use', 'resource_scale', 'resource_unit', 'storage_cap_equals', 'storage_cap_max', 'storage_cap_min', 'storage_cap_per_unit', 'storage_initial', 'storage_loss', 'units_equals', 'units_equals_systemwide', 'units_max', 'units_max_systemwide', 'units_min']
@@ -99,9 +99,9 @@ tech_groups:
             parent: null
         constraints:
             energy_prod: true
+            resource: inf
             resource_eff: 1.0
             resource_unit: power
-            resource: inf
     transmission:
         required_constraints: [['energy_cap_equals', 'energy_cap_max', 'energy_cap_per_unit']]
         allowed_constraints: ['energy_cap_equals', 'energy_cap_max', 'energy_cap_per_unit', 'energy_cap_scale', 'energy_con', 'energy_eff', 'energy_eff_per_distance', 'energy_prod', 'lifetime', 'one_way']
