### 2018-06-13 5b0cb2e
Alphabetically sort `model` and `defaults` config YAML files (#122)
