### 2018-06-14 e88c213
Add function plot_energy_flows
Adding a new plot function showing the energy produced in locs, aside
with the energy transmitted by transmission lines. Energy is ordered by
techs. A slider is used to navigate through timesteps.
