+++ calliope/analysis/plotting/energy_flows.py
@@ -1,0 +1,289 @@
+"""
+Copyright (C) 2013-2018 Calliope contributors listed in AUTHORS.
+Licensed under the Apache 2.0 License (see LICENSE file).
+
+energy_flows.py
+~~~~~~~~~~~~~~~
+
+Plot energy flows data.
+
+"""
+
+import pandas as pd
+
+
+def _format_date(date, format):
+    ts = pd.to_datetime(date)
+    return ts.strftime(format)
+
+def _scale_prod_factor(carrier_prod):
+    return 100/abs(carrier_prod.values.max()-carrier_prod.values.min())
+
+
+def _line(loc_coordinates, location, carrier, tech, prod, scale_factor,
+            techs_colors, is_initial_timestep):
+    [transmission_type, from_location] = tech.split(":")
+    hover_info = "%s>%s: %.2f by %s (%s)" % \
+        (from_location, location, prod, transmission_type, carrier)
+    # ex: "Region1>Region2: 256.54 by gas_transmission (gas)"
+
+    line = dict(
+        visible=False,
+        mode="lines",
+        hoverinfo="text",
+        text="",
+        line=dict(
+            width=prod*scale_factor + 1,
+            color=techs_colors[transmission_type]
+        ),
+        legendgroup=transmission_type,
+        name=tech,
+        opacity=0.6,
+    )
+
+    line_info_marker = dict(
+        visible=False,
+        mode="markers",
+        hoverinfo="text",
+        text=hover_info,
+        marker=dict(
+            symbol="square",
+            opacity=0,
+            color=techs_colors[transmission_type]
+        ),
+        legendgroup=False,
+        name=""
+    )
+
+    h_coord, v_coord = "lon", "lat" # by default
+    if set(loc_coordinates.index) == set(["x", "y"]):
+        h_coord, v_coord = "x", "y"
+        line["type"] = "scatter"
+        line_info_marker["type"] = "scatter"
+    elif set(loc_coordinates.index) == set(["lon", "lat"]):
+        h_coord, v_coord = "lon", "lat"
+        line["type"] = "scattergeo"
+        line_info_marker["type"] = "scattergeo"
+
+    line[h_coord]= [
+        loc_coordinates[from_location][h_coord],
+        loc_coordinates[location][h_coord]
+    ]
+    line[v_coord]= [
+        loc_coordinates[from_location][v_coord],
+        loc_coordinates[location][v_coord]
+    ]
+    line_info_marker[h_coord] = [ (1/2) * (loc_coordinates[from_location][h_coord] + \
+        loc_coordinates[location][h_coord]) ]
+    line_info_marker[v_coord] = [ (1/2) * (loc_coordinates[from_location][v_coord] + \
+        loc_coordinates[location][v_coord]) ]
+
+    if is_initial_timestep:
+    # plot only the first timestep data when the chart is initialized
+        line["visible"]=True
+        line_info_marker["visible"]=True
+
+
+    return [line, line_info_marker]
+
+
+def _marker(loc_coordinates, location, carrier, tech, prod, scale_factor,
+            techs_colors, is_initial_timestep):
+    hover_info = "%s: %.2f of %s (%s)" % \
+        (location, prod, tech, carrier)
+    # ex: "Region1: 3552.65 of pipe_import (gas)"
+
+    marker_dict = dict(
+        visible=False,
+        hoverinfo="text",
+        text=hover_info,
+        mode="markers",
+        marker=dict(
+            symbol="dot",
+            opacity=0.6,
+            size=prod*scale_factor + 1,
+            color=techs_colors[tech],
+        ),
+        legendgroup=tech,
[…]
