### 2018-06-16 94d5727
Add `excess_supply` variable when ensuring feasibility
