+++ calliope/backend/pyomo/variables.py
@@ -92,4 +92,5 @@ def initialize_decision_variables(backend_model):
 
     if model_data_dict['attrs'].get('run.ensure_feasibility', False):
         backend_model.unmet_demand = po.Var(backend_model.loc_carriers, backend_model.timesteps, within=po.NonNegativeReals)
+        backend_model.excess_supply = po.Var(backend_model.loc_carriers, backend_model.timesteps, within=po.NegativeReals)
         backend_model.bigM = model_data_dict['attrs'].get('run.bigM')
