+++ changelog.rst
@@ -6,6 +6,8 @@ Release History
 0.6.3-dev
 ---------
 
+|fixed| When setting `ensure_feasibility`, the resulting `unmet_demand` variable can also be negative, accounting for possible infeasibility when there is excess supply, relative to demand (assuming no load shedding abilities). This is particularly pertinent when the `force_resource` constraint is in place.
+
 |fixed| When applying systemwide constraints to transmission technologies, they are no longer silently ignored. Instead, the constraint value is doubled (to account for the constant existence of a pair of technologies to describe one link) and applied to the relevant transmission techs.
 
 |new| ``calliope generate_runs`` in the command line interface can now produce scripts for remote clusters which require SLURM-based submission (``sbatch...``).
