### 2018-06-18 7cd60f7
removing line_marker_info from legend + linter errors fix
