+++ calliope/analysis/plotting/energy_flows.py
@@ -12,16 +12,17 @@ Plot energy flows data.
 import pandas as pd
 
 
-def _format_date(date, format):
+def _format_date(date, timeseries_format):
     ts = pd.to_datetime(date)
-    return ts.strftime(format)
+    return ts.strftime(timeseries_format)
+
 
 def _scale_prod_factor(carrier_prod):
-    return 100/abs(carrier_prod.values.max()-carrier_prod.values.min())
+    return 100 / abs(carrier_prod.values.max() - carrier_prod.values.min())
 
 
-def _line(loc_coordinates, location, carrier, tech, prod, scale_factor,
-            techs_colors, is_initial_timestep):
+def _line(loc_coordinates, location, carrier, tech, prod,
+          scale_factor, techs_colors, is_initial_timestep):
     [transmission_type, from_location] = tech.split(":")
     hover_info = "%s>%s: %.2f by %s (%s)" % \
         (from_location, location, prod, transmission_type, carrier)
@@ -33,7 +34,7 @@ def _line(loc_coordinates, location, carrier, tech, prod, scale_factor,
         hoverinfo="text",
         text="",
         line=dict(
-            width=prod*scale_factor + 1,
+            width=prod * scale_factor + 1,
             color=techs_colors[transmission_type]
         ),
         legendgroup=transmission_type,
@@ -51,11 +52,12 @@ def _line(loc_coordinates, location, carrier, tech, prod, scale_factor,
             opacity=0,
             color=techs_colors[transmission_type]
         ),
-        legendgroup=False,
-        name=""
+        legendgroup=transmission_type,
+        name=tech,
+        showlegend=False
     )
 
-    h_coord, v_coord = "lon", "lat" # by default
+    h_coord, v_coord = "lon", "lat"  # by default
     if set(loc_coordinates.index) == set(["x", "y"]):
         h_coord, v_coord = "x", "y"
         line["type"] = "scatter"
@@ -65,24 +67,23 @@ def _line(loc_coordinates, location, carrier, tech, prod, scale_factor,
         line["type"] = "scattergeo"
         line_info_marker["type"] = "scattergeo"
 
-    line[h_coord]= [
+    line[h_coord] = [
         loc_coordinates[from_location][h_coord],
         loc_coordinates[location][h_coord]
     ]
-    line[v_coord]= [
+    line[v_coord] = [
         loc_coordinates[from_location][v_coord],
         loc_coordinates[location][v_coord]
     ]
-    line_info_marker[h_coord] = [ (1/2) * (loc_coordinates[from_location][h_coord] + \
-        loc_coordinates[location][h_coord]) ]
-    line_info_marker[v_coord] = [ (1/2) * (loc_coordinates[from_location][v_coord] + \
-        loc_coordinates[location][v_coord]) ]
+    line_info_marker[h_coord] = [(1 / 2) * (loc_coordinates[from_location][h_coord] +
+                                 loc_coordinates[location][h_coord])]
+    line_info_marker[v_coord] = [(1 / 2) * (loc_coordinates[from_location][v_coord] +
+                                 loc_coordinates[location][v_coord])]
 
     if is_initial_timestep:
-    # plot only the first timestep data when the chart is initialized
-        line["visible"]=True
-        line_info_marker["visible"]=True
-
+        # plot only the first timestep data when the chart is initialized
+        line["visible"] = True
+        line_info_marker["visible"] = True
 
     return [line, line_info_marker]
 
@@ -101,14 +102,14 @@ def _marker(loc_coordinates, location, carrier, tech, prod, scale_factor,
         marker=dict(
             symbol="dot",
             opacity=0.6,
-            size=prod*scale_factor + 1,
[…]
