### 2018-06-18 a430121
adding HTML string test for energy_flows plotting function
