### 2018-06-20 d3c86d7
Docs: YAML comments no longer lost in conversion
