### 2018-06-22 2d0d704
Fixed length (18) for legend names and fixed size for legend markers
