### 2018-06-26 eab0dcb
``out_file`` -> ``to_file`` in summary plot docs
