### 2018-06-26 fcc3000
Update location of energy_flows helper functions; change legend solution
* Shared helper function with other plotting functions are now in util.py

* Return to Open Sans font, but force a plot margin that always ensures the plot is static

* Only have top level tech names in the legend (and move to their longhand names)
