+++ calliope/analysis/plotting/energy_flows.py
@@ -10,69 +10,15 @@ Plot energy flows data.
 
 import pandas as pd
 
-
-def _format_name(name, length):
-    if len(name) < length:
-        return _min_size_name(name, length) + " " * 3
-        # we add 3 spaces because break_name
-        # adds 3 points on too long names
-        # -> side effect but whatever, those functions won't be
-        # imported elsewhere
-    else:
-        return _break_name(name, length)
+from calliope.analysis.plotting.util import break_name, get_range
 
 
-def _break_name(name, length):
-    """Take a long technology name string and break it across multiple lines
-    of lenght 'length' """
-    name_breaks = len(name) / length
-    if name_breaks > 1:
-        initial_name = name
-        break_buffer = 0
-        for name_break in range(1, int(name_breaks) + 1):
-            # preferably break at a space
-            breakpoint = initial_name.rfind(
-                ' ', name_break * length - 10, name_break * length)
-            # -1 means rfind failed to find any space
-            if breakpoint != -1:
-                breakpoint += break_buffer
-                name = name[:breakpoint].rstrip() + '<br>' + \
-                    name[breakpoint:].lstrip()
-                break_buffer += 4
-            else:
-                breakpoint = length * name_break + break_buffer
-                name = name[:breakpoint].rstrip() + '...<br>' + \
-                    name[breakpoint:].lstrip()
-                # '...' is len 3
-                break_buffer += 7
-
-    return name
-
-
-def _min_size_name(name, length):
-    """ Take a tech name and a length and return the name string
-    with at least the length, adding " " when too short """
-    if len(name) < length:
-        return name + " " * (length - len(name))
-    else:
-        return name
-
-
-def _format_date(date, timeseries_format):
-    ts = pd.to_datetime(date)
-    return ts.strftime(timeseries_format)
-
-
-def _scale_prod_factor(carrier_prod):
-    return 100 / abs(carrier_prod.values.max() - carrier_prod.values.min())
-
-
-def _line(loc_coordinates, location, carrier, tech, prod,
-          scale_factor, techs_colors, is_initial_timestep):
+def _line(loc_coordinates, location, carrier, tech, prod, scale_factor,
+          techs_colors, is_initial_timestep, add_legend, name):
     [transmission_type, from_location] = tech.split(":")
     hover_info = "%s>%s: %.2f by %s (%s)" % \
         (from_location, location, prod, transmission_type, carrier)
-    # ex: "Region1>Region2: 256.54 by gas_transmission (gas)"
+    # e.g. "Region1>Region2: 256.54 by gas_transmission (gas)"
 
     line = dict(
         visible=False,
@@ -84,7 +30,6 @@ def _line(loc_coordinates, location, carrier, tech, prod,
             color=techs_colors[transmission_type]
         ),
         legendgroup=transmission_type,
-        name=_format_name(tech, 18),
         opacity=0.6,
         showlegend=False
     )
@@ -99,7 +44,7 @@ def _line(loc_coordinates, location, carrier, tech, prod,
             color=techs_colors[transmission_type]
         ),
         legendgroup=transmission_type,
-        name=_format_name(tech, 18),
+        name=break_name(name, 18),
         opacity=0.6,
     )
 
@@ -118,17 +63,10 @@ def _line(loc_coordinates, location, carrier, tech, prod,
         showlegend=False
     )
 
-    h_coord, v_coord = "lon", "lat"  # by default
     if set(loc_coordinates.index) == set(["x", "y"]):
         h_coord, v_coord = "x", "y"
-        line["type"] = "scatter"
-        line_legend["type"] = "scatter"
-        line_info_marker["type"] = "scatter"
     elif set(loc_coordinates.index) == set(["lon", "lat"]):
         h_coord, v_coord = "lon", "lat"
[…]
