### 2018-06-27 df65948
Update following review, inc. `excess_supply` -> `unused_supply`
