+++ calliope/analysis/postprocess.py
@@ -183,14 +183,14 @@ def clean_results(results, zero_threshold, timings):
         comment='Postprocessing: ' + comment
     )
 
-    # Combine excess_supply and unmet_demand into one variable
+    # Combine unused_supply and unmet_demand into one variable
     if ('unmet_demand' in results.data_vars.keys() or
-            'excess_supply' in results.data_vars.keys()):
+            'unused_supply' in results.data_vars.keys()):
         results['unmet_demand'] = (
-            results.get('unmet_demand', 0) + results.get('excess_supply', 0)
+            results.get('unmet_demand', 0) + results.get('unused_supply', 0)
         )
 
-        results = results.drop('excess_supply')
+        results = results.drop('unused_supply')
 
         if not results.unmet_demand.sum():
 
