+++ calliope/backend/pyomo/objective.py
@@ -31,10 +31,10 @@ def minmax_cost_optimization(backend_model, cost_class, sense):
 
     """
     def obj_rule(backend_model):
-        if hasattr(backend_model, 'unmet_demand'):
+        if backend_model.__calliope_model_data__['attrs'].get('run.ensure_feasibility', False):
             unmet_demand = sum(
                 backend_model.unmet_demand[loc_carrier, timestep] -
-                backend_model.excess_supply[loc_carrier, timestep]
+                backend_model.unused_supply[loc_carrier, timestep]
                 for loc_carrier in backend_model.loc_carriers
                 for timestep in backend_model.timesteps
             ) * backend_model.bigM
