+++ calliope/backend/pyomo/constraints/energy_balance.py
@@ -106,16 +106,16 @@ def system_balance_constraint_rule(backend_model, loc_carrier, timestep):
 
     """
     prod, con, export = get_loc_tech_carriers(backend_model, loc_carrier)
-    if hasattr(backend_model, 'bigM'):
+    if backend_model.__calliope_model_data__['attrs'].get('run.ensure_feasibility', False):
         unmet_demand = backend_model.unmet_demand[loc_carrier, timestep]
-        excess_supply = backend_model.excess_supply[loc_carrier, timestep]
+        unused_supply = backend_model.unused_supply[loc_carrier, timestep]
     else:
-        unmet_demand = excess_supply = 0
+        unmet_demand = unused_supply = 0
 
     backend_model.system_balance[loc_carrier, timestep].expr = (
         sum(backend_model.carrier_prod[loc_tech_carrier, timestep] for loc_tech_carrier in prod) +
         sum(backend_model.carrier_con[loc_tech_carrier, timestep] for loc_tech_carrier in con) +
-        unmet_demand + excess_supply
+        unmet_demand + unused_supply
     )
 
     return backend_model.system_balance[loc_carrier, timestep] == 0
