### 2018-06-27 9582785
renaming energy_flows() function to flows() + some cleaning
