+++ calliope/analysis/plotting/plotting.py
@@ -19,7 +19,7 @@ from calliope.exceptions import warn
 from calliope.analysis.plotting.capacity import plot_capacity
 from calliope.analysis.plotting.timeseries import plot_timeseries
 from calliope.analysis.plotting.transmission import plot_transmission
-from calliope.analysis.plotting.energy_flows import plot_energy_flows
+from calliope.analysis.plotting.flows import plot_flows
 from calliope.analysis.plotting.util import type_of_script
 
 
@@ -192,12 +192,12 @@ class ModelPlotMethods:
 
     transmission.__doc__ = plot_transmission.__doc__.rstrip() + _docstring_additions
 
-    def energy_flows(self, **kwargs):
+    def flows(self, **kwargs):
         self.check_optimality()
-        data, layout = plot_energy_flows(self._model, **kwargs)
+        data, layout = plot_flows(self._model, **kwargs)
         return _plot(data, layout, **kwargs)
 
-    energy_flows.__doc__ = plot_energy_flows.__doc__.rstrip() + _docstring_additions
+    flows.__doc__ = plot_flows.__doc__.rstrip() + _docstring_additions
 
     def summary(self, **kwargs):
         self.check_optimality()
