+++ calliope/analysis/plotting/__init__.py
@@ -2,4 +2,4 @@ from calliope.analysis.plotting.plotting import plot_summary, ModelPlotMethods
 from calliope.analysis.plotting.capacity import plot_capacity
 from calliope.analysis.plotting.timeseries import plot_timeseries
 from calliope.analysis.plotting.transmission import plot_transmission
-from calliope.analysis.plotting.energy_flows import plot_energy_flows
+from calliope.analysis.plotting.flows import plot_flows
