### 2018-06-28 ac6443e
Add `unused_supply` variable when ensuring feasibility (#124)
