+++ calliope/test/test_constraint_results.py
@@ -2,6 +2,7 @@ import pytest
 from pytest import approx
 
 import calliope
+from calliope.test.common.util import build_test_model as build_model
 
 _OVERRIDE_FILE = calliope.examples._PATHS['national_scale'] + '/overrides.yaml'
 
@@ -68,3 +69,84 @@ class TestNationalScaleExampleModelSenseChecks:
         # assert constraint_string in model._backend_model.reserve_margin_constraint.pprint()
 
         assert float(model.results.cost.sum()) == approx(282487.35489)
+
+
+class TestModelSettings:
+    def test_feasibility(self):
+
+        def override(feasibility, cap_val):
+            override_dict = {
+                'locations.0.techs': {'test_supply_elec': {}, 'test_demand_elec': {}},
+                'links.0,1.exists': False,
+                # pick a time subset where demand is uniformally -10 throughout
+                'model.subset_time': ['2005-01-01 06:00:00', '2005-01-01 08:00:00'],
+                'run.ensure_feasibility': feasibility, 'run.bigM': 1e3,
+                # Allow setting resource and energy_cap_max/equals to force infeasibility
+                'techs.test_supply_elec.constraints': {
+                    'resource': cap_val, 'energy_eff': 1, 'energy_cap_equals': 15,
+                    'force_resource': True
+                }
+            }
+
+            return override_dict
+
+        # Feasible case, unmet_demand/unused_supply is deleted
+        model_10 = build_model(
+            override_dict=override(True, 10),
+            override_groups='investment_costs'
+        )
+        model_10.run()
+        for i in ['unmet_demand', 'unused_supply']:
+            assert hasattr(model_10._backend_model, i)
+            assert i not in model_10._model_data.data_vars.keys()
+
+        # Infeasible case, unmet_demand is required
+        model_5 = build_model(
+            override_dict=override(True, 5),
+            override_groups='investment_costs'
+        )
+        model_5.run()
+        assert hasattr(model_5._backend_model, 'unmet_demand')
+        assert hasattr(model_5._backend_model, 'unused_supply')
+        assert model_5._model_data['unmet_demand'].sum() == 15
+        assert 'unused_supply' not in model_5._model_data.data_vars.keys()
+
+        # Infeasible case, unused_supply is required
+        model_15 = build_model(
+            override_dict=override(True, 15),
+            override_groups='investment_costs'
+        )
+        model_15.run()
+        assert hasattr(model_15._backend_model, 'unmet_demand')
+        assert hasattr(model_15._backend_model, 'unused_supply')
+        assert model_15._model_data['unmet_demand'].sum() == -15
+        assert 'unused_supply' not in model_15._model_data.data_vars.keys()
+
+        assert (
+            model_15._backend_model.obj.expr() - model_10._backend_model.obj.expr() ==
+            approx(1e3 * 15)
+        )
+
+        assert (
+            model_5._backend_model.obj.expr() - model_10._backend_model.obj.expr() ==
+            approx(1e3 * 15)
+        )
+
+        # Infeasible cases = non-optimal termination
+        # too much supply
+        model = build_model(
+            override_dict=override(False, 15),
+            override_groups='investment_costs'
+        )
+        model.run()
+        assert not hasattr(model._backend_model, 'unmet_demand')
+        assert not hasattr(model._backend_model, 'unused_supply')
+        assert not model._model_data.attrs['termination_condition'] == 'optimal'
+
+        # too little supply
+        model = build_model(
+            override_dict=override(False, 5),
+            override_groups='investment_costs'
+        )
+        model.run()
+        assert not model._model_data.attrs['termination_condition'] == 'optimal'
