### 2018-06-29 #712 @katyhuff
Hi @mdoucet and @gonsie -- - It seems to me you would be excellent reviewers for this submission. Are you willing to be assigned to review this software paper submission? We'll need at least two volunteers, but the more the merrier. Information about being a reviewer can be found here : http://joss.theoj.org/about#reviewer_guidelines .
