### 2018-07-10 a78dae1
Fix national scale operate mode & test it
