+++ calliope/backend/checks.py
@@ -11,6 +11,7 @@ Checks for model consistency and possible errors when preparing run in the backe
 import ruamel.yaml
 
 import numpy as np
+import pandas as pd
 import xarray as xr
 from calliope.core.attrdict import AttrDict
 
@@ -37,7 +38,7 @@ def check_operate_params(model_data):
     comments = AttrDict()
 
     def _get_param(loc_tech, var):
-        if _is_in(loc_tech, var) and not np.isnan(model_data[var].loc[loc_tech].item()):
+        if _is_in(loc_tech, var) and not pd.isnull(model_data[var].loc[loc_tech].item()):
             param = model_data[var].loc[loc_tech].item()
         else:
             param = defaults[var]
@@ -56,7 +57,7 @@ def check_operate_params(model_data):
     for loc_tech in model_data.loc_techs.values:
         energy_cap = model_data.energy_cap.loc[loc_tech].item()
         # Must have energy_cap defined for all relevant techs in the model
-        if (np.isnan(energy_cap) or np.isinf(energy_cap)) and not _is_in(loc_tech, 'force_resource'):
+        if (pd.isnull(energy_cap) or np.isinf(energy_cap)) and not _is_in(loc_tech, 'force_resource'):
             errors.append(
                 'Operate mode: User must define a finite energy_cap (via '
                 'energy_cap_equals or energy_cap_max) for {}'.format(loc_tech)
@@ -65,21 +66,22 @@ def check_operate_params(model_data):
         elif _is_in(loc_tech, 'loc_techs_finite_resource'):
             # force resource overrides capacity constraints, so set capacity constraints to infinity
             if _is_in(loc_tech, 'force_resource'):
-                energy_cap = model_data.energy_cap.loc[loc_tech] = np.inf
-                warnings.append(
-                    'Energy capacity constraint removed from {} as force_resource '
-                    'is applied'.format(loc_tech)
-                )
+                if not _is_in(loc_tech, 'loc_techs_store'):
+                    energy_cap = model_data.energy_cap.loc[loc_tech] = np.inf
+                    warnings.append(
+                        'Energy capacity constraint removed from {} as '
+                        'force_resource is applied'.format(loc_tech)
+                    )
                 if _is_in(loc_tech, 'resource_cap'):
                     resource_cap = model_data.resource_cap.loc[loc_tech] = np.inf
                     warnings.append(
-                        'Resource capacity constraint removed from {} as force_resource '
-                        'is applied'.format(loc_tech)
+                        'Resource capacity constraint removed from {} as '
+                        'force_resource is applied'.format(loc_tech)
                     )
             # Cannot have infinite resource area (physically impossible)
             if _is_in(loc_tech, 'loc_techs_area'):
                 area = model_data.resource_area.loc[loc_tech].item()
-                if np.isnan(area) or np.isinf(area):
+                if pd.isnull(area) or np.isinf(area):
                     errors.append(
                         'Operate mode: User must define a finite resource_area '
                         '(via resource_area_equals or resource_area_max) for {}, '
@@ -93,7 +95,6 @@ def check_operate_params(model_data):
                 energy_cap_scale = _get_param(loc_tech, 'energy_cap_scale')
                 resource_eff = _get_param(loc_tech, 'resource_eff')
                 energy_eff = _get_param(loc_tech, 'energy_eff')
-                energy_cap_scale = _get_param(loc_tech, 'energy_cap_scale')
                 resource = model_data.resource.loc[loc_tech].values
                 if (energy_cap is not None and
                     any(resource * resource_scale * resource_eff >
@@ -120,7 +121,7 @@ def check_operate_params(model_data):
             if _is_in(loc_tech, 'charge_rate'):
[…]
