### 2018-07-16 #825 @whedon
[REVIEW]: Calliope: a multi-scale energy systems modelling framework
**Submitting author:** @sjpfenninger (<a href="http://orcid.org/0000-0002-8420-9498">Stefan Pfenninger</a>)
**Repository:** <a href="https://github.com/calliope-project/calliope" target ="_blank">https://github.com/calliope-project/calliope</a>
**Version:** v0.6.2
