### 2018-07-17 calliope-project/calliope#129 @mdoucet
Problem with installation
## Problem description

Hi- This is in reference to your [JOSS review](https://github.com/openjournals/joss-reviews/issues/825).

I tried following your [installation instructions](https://calliope.readthedocs.io/en/stable/user/installation.html#recommended-installation-method). The installation process below proceeded just fine, but the installed software didn't work.

## Steps to reproduce the problem

After this:
``$ conda create -c conda-forge -n calliope python=3.6 calliope``

I got this:

```
$ python
Python 3.6.5 | packaged by conda-forge | (default, Apr  6 2018, 13:44:09)
[GCC 4.2.1 Compatible Apple LLVM 6.1.0 (clang-602.0.53)] on darwin
>>> import calliope
Traceback (most recent call last):
  File "/.../anaconda/envs/calliope/lib/python3.6/site-packages/numpy/core/__init__.py", line 16, in <module>
    from . import multiarray
ImportError: dlopen(/.../anaconda/envs/calliope/lib/python3.6/site-packages/numpy/core/multiarray.cpython-36m-darwin.so, 2): Library not loaded: @rpath/libgfortran.3.dylib
  Referenced from: /.../anaconda/envs/calliope/lib/libopenblas.dylib
  Reason: image not found

Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/.../anaconda/envs/calliope/lib/python3.6/site-packages/calliope/__init__.py", line 2, in <module>
    from calliope.core import AttrDict, Model, read_netcdf, _logger, set_log_level
  File "/.../anaconda/envs/calliope/lib/python3.6/site-packages/calliope/core/__init__.py", line 1, in <module>
    from calliope.core.model import Model, read_netcdf
  File "/.../anaconda/envs/calliope/lib/python3.6/site-packages/calliope/core/model.py", line 12, in <module>
    import numpy as np
  File "/.../anaconda/envs/calliope/lib/python3.6/site-packages/numpy/__init__.py", line 142, in <module>
    from . import add_newdocs
  File "/.../anaconda/envs/calliope/lib/python3.6/site-packages/numpy/add_newdocs.py", line 13, in <module>
    from numpy.lib import add_newdoc
  File "/.../anaconda/envs/calliope/lib/python3.6/site-packages/numpy/lib/__init__.py", line 8, in <module>
    from .type_check import *
  File "/.../anaconda/envs/calliope/lib/python3.6/site-packages/numpy/lib/type_check.py", line 11, in <module>
    import numpy.core.numeric as _nx
  File "/.../anaconda/envs/calliope/lib/python3.6/site-packages/numpy/core/__init__.py", line 26, in <module>
    raise ImportError(msg)
ImportError:
Importing the multiarray numpy extension module failed.  Most
likely you are trying to import a failed build of numpy.
If you're working with a numpy git repo, try `git clean -xdf` (removes all
files not under version control).  Otherwise reinstall numpy.

Original error was: dlopen(/.../anaconda/envs/calliope/lib/python3.6/site-packages/numpy/core/multiarray.cpython-36m-darwin.so, 2): Library not loaded: @rpath/libgfortran.3.dylib
  Referenced from: /.../anaconda/envs/calliope/lib/libopenblas.dylib
  Reason: image not found
```

Interestingly, I could just create myself an environment and build calliope by hand and that worked fine. I could run through your first tutorial.

So the following worked fine:
```
$ conda env create -n calliope --file requirements/base.yml
$ python setup.py install
```

## Calliope version

0.6.2-1
