### 2018-07-24 calliope-project/calliope#129 @sjpfenninger
@mdoucet Hmm, I cannot reproduce this, and it looks like its actually an issue with numpy installed through `conda` not finding some underlying library, so it seems strange that your second installation account worked (as it should have installed the same libraries anyway). It looks like you're on macOS, what version exactly? If you re-run `conda create -c conda-forge -n calliope python=3.6 calliope` (perhaps with `-n calliope_test` to create a separate test environment), do you then get the same error again?
