### 2018-07-24 calliope-project/calliope#129 @mdoucet
OK. Apparently I had a strange development version of ```conda``` installed. After updating to the latest version and trying again it worked fine. I also tried on one of our RHEL7 machines just for fun and that worked to. So it's all good.
