### 2018-07-27 #825 @mdoucet
I've completed my review. While I'm not a subject matter expert of energy systems modeling, I found the documentation quite clear. I was able to go through examples without any issue. The code itself is well organized, with great use of continuous integration and automated testing.
