### 2018-07-29 #825 @ecotillasanchez
It was a pleasure to review this work by Pfenninger and Pickering. Excellent software and documentation. Please find below some suggestions:

* It is clear from the paper and the examples that the analysis scales well from urban to national networks. For a power systems user, however, it would be good to clarify that building a transmission model with Calliope is different than a traditional power flow analysis (Ybus, generation, demand -> dc or ac power flow equations -> voltage magnitude and angles).

* Similarly, perhaps note that the focus is power system planning (as opposed to operations).

* The tutorials cover a great amount of content, options, and optimization constraints. For a given network size, I would suggest a performance comparison between some of the external solvers, or simply some pointers to help users decide before finding possible roadblocks.
