### 2018-07-30 13548a0
Fixed random ordering when updating `storage_initial`
