+++ calliope/analysis/plotting/flows.py
@@ -13,12 +13,11 @@ import pandas as pd
 from calliope.analysis.plotting.util import break_name, get_range
 
 
-def _line(loc_coordinates, location, carrier, tech, prod, scale_factor,
+def _line(locs_coordinates, transmission_type, to_location, from_location, carrier, tech, prod, scale_factor,
           techs_colors, is_initial_timestep, add_legend, name):
-    [transmission_type, from_location] = tech.split(":")
     # e.g. "Region1->Region2: 256.54 by gas_transmission (gas)"
     hover_info = "%s->%s: %.2f by %s (%s)" % \
-        (from_location, location, prod, transmission_type, carrier)
+        (from_location, to_location, prod, transmission_type, carrier)
 
     line = dict(
         visible=False,
@@ -63,25 +62,25 @@ def _line(loc_coordinates, location, carrier, tech, prod, scale_factor,
         showlegend=False
     )
 
-    if set(loc_coordinates.index) == set(["x", "y"]):
+    if set(locs_coordinates.index) == set(["x", "y"]):
         h_coord, v_coord = "x", "y"
-    elif set(loc_coordinates.index) == set(["lon", "lat"]):
+    elif set(locs_coordinates.index) == set(["lon", "lat"]):
         h_coord, v_coord = "lon", "lat"
 
     line[h_coord] = [
-        loc_coordinates[from_location][h_coord],
-        loc_coordinates[location][h_coord]
+        locs_coordinates[from_location][h_coord],
+        locs_coordinates[to_location][h_coord]
     ]
     line[v_coord] = [
-        loc_coordinates[from_location][v_coord],
-        loc_coordinates[location][v_coord]
+        locs_coordinates[from_location][v_coord],
+        locs_coordinates[to_location][v_coord]
     ]
     line_legend[h_coord] = [None]
     line_legend[v_coord] = [None]
-    line_info_marker[h_coord] = [(1 / 2) * (loc_coordinates[from_location][h_coord] +
-                                 loc_coordinates[location][h_coord])]
-    line_info_marker[v_coord] = [(1 / 2) * (loc_coordinates[from_location][v_coord] +
-                                 loc_coordinates[location][v_coord])]
+    line_info_marker[h_coord] = [(1 / 2) * (locs_coordinates[from_location][h_coord] +
+                                 locs_coordinates[to_location][h_coord])]
+    line_info_marker[v_coord] = [(1 / 2) * (locs_coordinates[from_location][v_coord] +
+                                 locs_coordinates[to_location][v_coord])]
 
     if is_initial_timestep:
         # plot only the first timestep data when the chart is initialized
@@ -94,7 +93,7 @@ def _line(loc_coordinates, location, carrier, tech, prod, scale_factor,
         return [line, line_info_marker]
 
 
-def _marker(loc_coordinates, location, carrier, tech, prod, scale_factor,
+def _marker(locs_coordinates, location, carrier, tech, prod, scale_factor,
             techs_colors, is_initial_timestep, add_legend, name):
     # Example: "Region1: 3552.65 of pipe_import (gas)"
     hover_info = "%s: %.2f of %s (%s)" % \
@@ -130,13 +129,13 @@ def _marker(loc_coordinates, location, carrier, tech, prod, scale_factor,
         name=break_name(name, 18)
     )
 
-    if set(loc_coordinates.index) == set(["x", "y"]):
+    if set(locs_coordinates.index) == set(["x", "y"]):
         h_coord, v_coord = "x", "y"
-    elif set(loc_coordinates.index) == set(["lon", "lat"]):
+    elif set(locs_coordinates.index) == set(["lon", "lat"]):
[…]
