### 2018-08-03 73f4376
Cleaning and adding a line to changelog
