### 2018-08-03 d26c0d4
Fix capacity plotting issue (#130)
