### 2018-08-07 #825 @gonsie
Should the version number be updated to the latest release (v0.6.2)?
