### 2018-08-13 #825 @sjpfenninger
@gonsie Yes that would make sense. @jedbrown, should we modify the submission now or wait for all reviews to be completed?
