### 2018-08-13 #825 @sjpfenninger
I guess the paper also still says 0.6.0 - should I modify `paper.md` or do that at the very end of the review process?
