### 2018-08-14 4d40f7e
Update operate mode checks
