+++ test.yaml
@@ -1,124 +1,0 @@
-name: calliope_dev
-channels:
-  - anaconda
-  - conda-forge
-  - defaults
-dependencies:
-  - libgcc-ng=7.2.0=hdf63c60_3
-  - libgfortran=3.0.0=1
-  - libstdcxx-ng=7.2.0=hdf63c60_3
-  - appdirs=1.4.3=py_0
-  - asn1crypto=0.24.0=py36_0
-  - atomicwrites=1.1.5=py36_0
-  - attrs=18.1.0=py_0
-  - backcall=0.1.0=py_0
-  - blas=1.1=openblas
-  - bokeh=0.12.16=py36_0
-  - bottleneck=1.2.1=py36_1
-  - ca-certificates=2018.4.16=0
-  - certifi=2018.4.16=py36_0
-  - cffi=1.11.5=py36_0
-  - cftime=1.0.0=py36_0
-  - chardet=3.0.4=py36_0
-  - click=6.7=py_1
-  - cloudpickle=0.5.3=py_0
-  - colorlover=0.2.1=py_0
-  - coverage=4.5.1=py36_0
-  - cryptography=2.2.1=py36_0
-  - cufflinks-py=0.12.1=py_0
-  - curl=7.60.0=0
-  - cytoolz=0.9.0.1=py36_0
-  - dask=0.17.5=py_0
-  - dask-core=0.17.5=py_0
-  - decorator=4.3.0=py_0
-  - distributed=1.21.8=py36_0
-  - glpk=4.61=1
-  - gmp=6.1.2=0
-  - h5netcdf=0.5.1=py_0
-  - h5py=2.8.0=py36h470a237_0
-  - hdf4=4.2.13=0
-  - hdf5=1.10.1=2
-  - heapdict=1.0.0=py36_0
-  - idna=2.6=py36_1
-  - ipdb=0.11=py36_0
-  - ipython=6.4.0=py36_0
-  - ipython_genutils=0.2.0=py36_0
-  - jedi=0.12.0=py36_0
-  - jinja2=2.10=py36_0
-  - jpeg=9b=2
-  - krb5=1.14.6=0
-  - libffi=3.2.1=3
-  - libnetcdf=4.6.1=2
-  - libssh2=1.8.0=2
-  - locket=0.2.0=py36_1
-  - markupsafe=1.0=py36_0
-  - more-itertools=4.2.0=py_0
-  - msgpack-python=0.5.6=py36h2d50403_2
-  - natsort=5.3.2=py_0
-  - ncurses=5.9=10
-  - netcdf4=1.4.0=py36_0
-  - nose=1.3.7=py36_2
-  - numexpr=2.6.5=py36_0
-  - numpy=1.14.1=py36_blas_openblas_200
-  - openblas=0.2.20=8
-  - openssl=1.0.2o=0
-  - packaging=17.1=py_0
-  - pandas=0.22.0=py36_1
-  - parso=0.2.1=py_0
-  - partd=0.3.8=py36_0
-  - pexpect=4.6.0=py36_0
-  - pickleshare=0.7.4=py36_0
-  - pip=9.0.3=py36_0
-  - plotly=2.5.1=py36_0
-  - pluggy=0.6.0=py_0
-  - ply=3.11=py36_0
-  - prompt_toolkit=1.0.15=py36_0
-  - psutil=5.4.5=py36_0
-  - ptyprocess=0.5.2=py36_0
-  - py=1.5.3=py_0
-  - pycparser=2.18=py36_0
-  - pygments=2.2.0=py36_0
-  - pyomo=5.4.3=py36_0
-  - pyopenssl=18.0.0=py36_0
-  - pyparsing=2.2.0=py36_0
[…]
