### 2018-08-16 35a09a8
Add groups_file option to generate_runs CLI
