+++ calliope/core/util/generate_runs.py
@@ -10,31 +10,66 @@ in parallel on a cluster or sequentially on any machine.
 
 """
 
+import itertools
 import os
 
 import pandas as pd
 from calliope.core import AttrDict
 
 
-def generate_runs(model_file, override_file, groups=None, additional_args=None):
+def generate_runs(
+        model_file, override_file,
+        groups_file=None, groups=None, additional_args=None):
     """
-    Returns a list of "calliope run" invocations.
+    Returns a list of ``calliope run`` invocations.
 
-    groups specified as group1{,group2,...}{;group3...}
+    Groups can be specified as ``group1{,group2,...}{;group3...}``,
+    for example:
+        * ``group1,group2;group1,group3`` to generate two runs, one with
+          group1 and group2, one with group1 and group3
+        * ``group1;group2;group3`` to generate three runs with one group
+          each
 
-    if groups not specified, use all groups in the override_file, one by one
+    If neither ``groups_file`` nor ``groups`` specified,
+    uses all groups in the given ``override_file``, one by one.
 
     """
-    if groups is None:
+    if groups_file and groups:
+        raise ValueError(
+            'Only one of `groups_file` or `groups` may '
+            'be defined.'
+        )
+
+    if groups_file:
+        groups = AttrDict.from_yaml(groups_file)
+        if 'combinations' in groups and 'groups' in groups:
+            raise ValueError(
+                'Only one of `combinations` or `groups` may '
+                'be defined in the groups_file.'
+            )
+        if 'combinations' in groups:
+            combinations = list(itertools.product(*groups['combinations']))
+            runs = [','.join(i) for i in combinations]
+        elif 'groups' in groups:
+            runs = groups['groups']
+        else:
+            raise KeyError(
+                '{} defines neither combinations nor '
+                'groups.'.format(groups_file)
+            )
+    elif groups:
+        runs = groups.split(';')
+    else:  # Run for all groups
         overrides = AttrDict.from_yaml(override_file)
         runs = overrides.keys()
-    else:
-        runs = groups.split(';')
 
     commands = []
 
     for i, run in enumerate(runs):
-        cmd = 'calliope run {model} --override_file {override}:{groups} --save_netcdf out_{i}_{groups}.nc --save_plots plots_{i}_{groups}.html {other_options}'.format(
+        cmd = ('calliope run {model} --override_file {override}:{groups} '
+               '--save_netcdf out_{i}_{groups}.nc '
+               '--save_plots plots_{i}_{groups}.html '
+               '{other_options}').format(
             i=i + 1,
             model=model_file,
             override=override_file,
@@ -46,8 +81,12 @@ def generate_runs(model_file, override_file, groups=None, additional_args=None):
     return commands
 
 
-def generate_bash_script(out_file, model_file, override_file, groups, additional_args=None, **kwargs):
-    commands = generate_runs(model_file, override_file, groups, additional_args)
+def generate_bash_script(
+        out_file, model_file, override_file, groups_file, groups,
+        additional_args=None, **kwargs):
+
+    commands = generate_runs(
+        model_file, override_file, groups_file, groups, additional_args)
 
     base_string = '    {i}) {cmd} ;;\n'
     lines_start = [
@@ -70,7 +109,10 @@ def generate_bash_script(out_file, model_file, override_file, groups, additional
         '',
     ]
 
-    lines_all = lines_start + [base_string.format(i=i + 1, cmd=cmd) for i, cmd in enumerate(commands)] + lines_end
+    lines_all = lines_start + [
[…]
