### 2018-08-16 e70e114
Add function plot_flows (#123)
Adding a new plot function showing the energy produced in locs and energy transmitted by transmission lines. Energy is ordered by techs. A slider is used to navigate through timesteps.
