+++ calliope/analysis/plotting/flows.py
@@ -1,0 +1,376 @@
+"""
+Copyright (C) 2013-2018 Calliope contributors listed in AUTHORS.
+Licensed under the Apache 2.0 License (see LICENSE file).
+
+flows.py
+~~~~~~~~~~~~~~~
+
+Plot energy flows data.
+"""
+
+import pandas as pd
+
+from calliope.analysis.plotting.util import break_name, get_range
+
+
+def _line(locs_coordinates, transmission_type, to_location, from_location, carrier, tech, prod, scale_factor,
+          techs_colors, is_initial_timestep, add_legend, name):
+    # e.g. "Region1->Region2: 256.54 by gas_transmission (gas)"
+    hover_info = "%s->%s: %.2f by %s (%s)" % \
+        (from_location, to_location, prod, transmission_type, carrier)
+
+    line = dict(
+        visible=False,
+        mode="lines",
+        hoverinfo="text",
+        text="",
+        line=dict(
+            width=prod * scale_factor + 1,
+            color=techs_colors[transmission_type]
+        ),
+        legendgroup=transmission_type,
+        opacity=0.6,
+        showlegend=False
+    )
+
+    line_legend = dict(
+        visible=False,
+        mode="lines",
+        hoverinfo="text",
+        text="",
+        line=dict(
+            width=10,
+            color=techs_colors[transmission_type]
+        ),
+        legendgroup=transmission_type,
+        name=break_name(name, 18),
+        opacity=0.6,
+    )
+
+    line_info_marker = dict(
+        visible=False,
+        mode="markers",
+        hoverinfo="text",
+        text=hover_info,
+        marker=dict(
+            symbol="square",
+            opacity=0,
+            color=techs_colors[transmission_type]
+        ),
+        legendgroup=transmission_type,
+        name=tech,
+        showlegend=False
+    )
+
+    if set(locs_coordinates.index) == set(["x", "y"]):
+        h_coord, v_coord = "x", "y"
+    elif set(locs_coordinates.index) == set(["lon", "lat"]):
+        h_coord, v_coord = "lon", "lat"
+
+    line[h_coord] = [
+        locs_coordinates[from_location][h_coord],
+        locs_coordinates[to_location][h_coord]
+    ]
+    line[v_coord] = [
+        locs_coordinates[from_location][v_coord],
+        locs_coordinates[to_location][v_coord]
+    ]
+    line_legend[h_coord] = [None]
+    line_legend[v_coord] = [None]
+    line_info_marker[h_coord] = [(1 / 2) * (locs_coordinates[from_location][h_coord] +
+                                 locs_coordinates[to_location][h_coord])]
+    line_info_marker[v_coord] = [(1 / 2) * (locs_coordinates[from_location][v_coord] +
+                                 locs_coordinates[to_location][v_coord])]
+
+    if is_initial_timestep:
+        # plot only the first timestep data when the chart is initialized
+        line["visible"] = True
+        line_legend["visible"] = True
+        line_info_marker["visible"] = True
+    if add_legend:
+        return [line, line_legend, line_info_marker]
+    else:
+        return [line, line_info_marker]
+
+
+def _marker(locs_coordinates, location, carrier, tech, prod, scale_factor,
+            techs_colors, is_initial_timestep, add_legend, name):
+    # Example: "Region1: 3552.65 of pipe_import (gas)"
+    hover_info = "%s: %.2f of %s (%s)" % \
+        (location, prod, tech, carrier)
+
+    marker_dict = dict(
+        visible=False,
+        hoverinfo="text",
+        text=hover_info,
+        mode="markers",
+        marker=dict(
+            symbol="dot",
+            opacity=0.6,
+            size=prod * scale_factor + 1,
[…]
