### 2018-08-16 fe563f6
Update changelog.rst
