+++ changelog.rst
@@ -6,11 +6,13 @@ Release History
 0.6.3-dev
 ---------
 
+|new| Addition of ``flows`` plotting function. This shows production and how much they exchange with other locations. It also provides a slider in order to see flows' evolution through time.
+
+|new| ``calliope generate_runs`` in the command line interface can now produce scripts for remote clusters which require SLURM-based submission (``sbatch...``).
+
 |fixed| When setting `ensure_feasibility`, the resulting `unmet_demand` variable can also be negative, accounting for possible infeasibility when there is unused supply, once all demand has been met (assuming no load shedding abilities). This is particularly pertinent when the `force_resource` constraint is in place.
 
 |fixed| When applying systemwide constraints to transmission technologies, they are no longer silently ignored. Instead, the constraint value is doubled (to account for the constant existence of a pair of technologies to describe one link) and applied to the relevant transmission techs.
-
-|new| ``calliope generate_runs`` in the command line interface can now produce scripts for remote clusters which require SLURM-based submission (``sbatch...``).
 
 0.6.2 (2018-06-04)
 ------------------
@@ -26,8 +28,6 @@ Release History
 |new| CLI now uses the IPython debugger rather than built-in ``pdb``, which provides highlighting, tab completion, and other UI improvements
 
 |new| AttrDict now persists comments when reading from and writing to YAML files, and gains an API to view, add and remove comments on keys
-
-|new| Plotting flows() function. Shows how much energy locations produce and how much they exchange with other locations. Provided with a slider in order to see flows evolution through time. 
 
 |fixed| Fix CLI error when running a model without transmission technologies
 
