### 2018-08-17 11a0856
Fix national scale operate mode & test it (#128)
