+++ changelog.rst
@@ -12,6 +12,8 @@ Release History
 
 |new| ``calliope generate_runs`` in the command line interface can now produce scripts for remote clusters which require SLURM-based submission (``sbatch...``).
 
+|fixed| Permit groups in override files to specify imports of other YAML files
+
 0.6.2 (2018-06-04)
 ------------------
 
