### 2018-08-24 #825 @gonsie
I am unable the test the functionality of this due to issues with my work firewall (I am unable to install conda / dependencies). I may be able to review this on personal time (maybe another week or so), or I'm happy to defer to the other reviewers.
