### 2018-08-27 #825 @jedbrown
I notice the requirement `ruamel.yaml<=0.15` which would allow 0.15.0, but not 0.15.1. Is that really intentional? I also notice that ruamel.yaml 0.15.0 is not compatible with Python-3.7 and that Calliope is not compatible with current ruamel.yaml (at least because it now returns a type that implements the dict interface, but is not an instance of `dict`).

I think the reviews above are sufficient, but wanted to give an opportunity to address version compatibility issues because it would ease non-conda installation and coupling of Calliope with other software.
