### 2018-08-28 calliope-project/calliope#137 @sjpfenninger
Docs: more info on differences between solver options
## Problem description

Raised by @ecotillasanchez during our [JOSS review](https://github.com/openjournals/joss-reviews/issues/825):

> The tutorials cover a great amount of content, options, and optimization constraints. For a given network size, I would suggest a performance comparison between some of the external solvers, or simply some pointers to help users decide before finding possible roadblocks.
