### 2018-08-28 calliope-project/calliope#138 @sjpfenninger
Docs: detail on differences to power system modelling terminology
## Problem description

Raised by @ecotillasanchez during our [JOSS review](https://github.com/openjournals/joss-reviews/issues/825):

> It is clear from the paper and the examples that the analysis scales well from urban to national networks. For a power systems user, however, it would be good to clarify that building a transmission model with Calliope is different than a traditional power flow analysis (Ybus, generation, demand -> dc or ac power flow equations -> voltage magnitude and angles).

> Similarly, perhaps note that the focus is power system planning (as opposed to operations).
