### 2018-08-28 calliope-project/calliope#139 @sjpfenninger
Fix ruamel.yaml and Python 3.7 compatibility
## Problem description

Raised by @jedbrown during our [JOSS review](https://github.com/openjournals/joss-reviews/issues/825):

> I notice the requirement ruamel.yaml<=0.15 which would allow 0.15.0, but not 0.15.1. Is that really intentional? I also notice that ruamel.yaml 0.15.0 is not compatible with Python-3.7 and that Calliope is not compatible with current ruamel.yaml (at least because it now returns a type that implements the dict interface, but is not an instance of dict).

## Steps to address the problem

* Determine API changes in ruamel.yaml since 0.15
* Update AttrDict to work with >0.15
* Possibly simplify or remove our comment pass-through ability to reduce the complexity of AttrDict internals
