### 2018-08-28 #825 @sjpfenninger
Thanks @jedbrown for raising the `ruamel.yaml` requirement issue. The version pinning is based on the [ruamel.yaml README](https://pypi.org/project/ruamel.yaml/):

> Starting with version 0.15.0 the way YAML files are loaded and dumped is changing. See the API doc for details. Currently existing functionality will throw a warning before being changed/removed. For production systems you should pin the version being used with ``ruamel.yaml<=0.15``.

Not providing Python 3.7 compatibility is certainly issue. We will investigate this. I don't think there's a quick fix, as we'll have to make some changes to our internals.
