### 2018-08-28 calliope-project/calliope#139 @brynpickering
It's still recommended to pin ruamel.yaml to <=0.15.0 (https://pypi.org/project/ruamel.yaml/), so I guess that is why we've gone with it. Is there any reason for us to maintain compatibility with Python 3.7?
