### 2018-08-28 calliope-project/calliope#139 @jedbrown
@brynpickering I read that as a recommendation to pin the version ("for production systems") instead of using an open-ended `<0.16` that would assume interface stability for all 0.15.x releases. I think it's weird (and possibly a mistake) to recommend `<=0.15.0` in the context where they are talking about a big incompatible change from the 0.14 series. Presumably `==0.15.0` would make more sense.

Python 3.7 is already default in rolling release Linux distros, Homebrew, and Anaconda, and will be in discrete releases sometime in the near future (Fedora 29 in October). You may be able to delay a few months, but you'll likely get lots more requests in the coming months.
