### 2018-08-29 calliope-project/calliope#139 @GiorgioBalestrieri
Hi all, my two cents on this: as @jedbrown pointed out, it should be obvious that supporting the last Python release should always be a goal for an open source project.
Besides the fact that Pyhon 3.7 comes with several cool features (data classes, simplified interactive debugging...), let's not go back to "why should we support Python 3?", those positions did not age well.
This being said, I understand it might take a while and this has to be prioritized against other development efforts, but I  think it should definitely be on the roadmap, at least.
I've never contributed to Calliope (yet, hopefully) so I'm really just expressing an opinion here, I trust the devs to know what's best for the project.
