### 2018-08-29 calliope-project/calliope#139 @brynpickering
Hi both, thanks for your comments! I agree that we should aim to stay up to date as far as is possible, but it hasn't been a priority. I also think we can push it for longer in not being a priority, especially if a. it causes dependency issues (I e. We would be moving quicker than our existing dependencies), and b. We are recommending users install it in an isolated conda environment. I'm aware that we will update to it in due course!

On this particular issue, looks like ruamel.yaml hasn't updated its pinning recommendation in a long while. But, it was this recommendation that led us to our existing pinning many months ago. Whether or not it should be `==` or `<=` is now something of a moot point.

Before we look to update to the latest ruamel.yaml, would we be better depending on PyYaml? We lose roundtrip parsing of comments, but perhaps gain a more robust package?
