+++ calliope/core/time/funcs.py
@@ -20,7 +20,6 @@ from calliope.core.util.logging import logger
 from calliope.core.preprocess.lookup import lookup_clusters
 
 
-
 def get_daily_timesteps(data, check_uniformity=False):
     daily_timesteps = [
         data.timestep_resolution.loc[i].values
@@ -78,10 +77,10 @@ def normalized_copy(data):
 def _copy_non_t_vars(data0, data1):
     """Copies non-t-indexed variables from data0 into data1, then
     returns data1"""
-    non_t_vars = [v for v in data0.data_vars
-                  if 'timesteps' not in data0[v].dims]
-    # Manually copy over variables not in `t`. If we don't do this,
-    # these vars get polluted with a superfluous `t` dimension
+    non_t_vars = [varname for varname, vardata in data0.data_vars.items()
+                  if 'timesteps' not in vardata.dims]
+    # Manually copy over variables not in `timesteps`. If we don't do this,
+    # these vars get polluted with a superfluous `timesteps` dimension
     for v in non_t_vars:
         data1[v] = data0[v]
     return data1
@@ -96,6 +95,23 @@ def _combine_datasets(data0, data1):
     ]
 
     return data_new
+
+
+def _drop_timestep_vars(data, timesteps):
+    timeseries_data = data.copy(deep=True)
+    # Save all coordinates, to ensure they can be added back in after clustering
+    data_coords = data.copy().coords
+    del data_coords['timesteps']
+
+    if timesteps is not None:
+        timeseries_data = timeseries_data.loc[{'timesteps': timesteps}]
+
+    timeseries_data = timeseries_data.drop([
+        varname for varname, vardata in data.data_vars.items()
+        if 'timesteps' not in vardata.dims
+    ])
+
+    return timeseries_data, data_coords
 
 
 def apply_clustering(data, timesteps, clustering_func, how, normalize=True,
@@ -136,20 +152,11 @@ def apply_clustering(data, timesteps, clustering_func, how, normalize=True,
     daily_timesteps = get_daily_timesteps(data, check_uniformity=True)
     timesteps_per_day = len(daily_timesteps)
 
-    # Save all coordinates, to ensure they can be added back in after clustering
-    data_coords = data.copy().coords
-    del data_coords['timesteps']
-    # Only apply clustering function on subset of masked timesteps
-    if timesteps is None:
-        data_to_cluster = data
-    else:
-        data_to_cluster = data.loc[{'timesteps': timesteps}]
+    # get a copy of the dataset with only timeseries variables,
+    # and get all coordinates of the original dataset, to reinstate later
+    data_to_cluster, data_coords = _drop_timestep_vars(data, timesteps)
 
-    # remove all variables that are not indexed over time
-    data_to_cluster = data_to_cluster.drop([
-        i for i in data.variables
-        if 'timesteps' not in data[i].dims or 'timestep_' in i
-    ])
+    data_to_cluster = data_to_cluster.drop(['timestep_weights', 'timestep_resolution'])
 
     for dim in data_to_cluster.dims:
         data_to_cluster[dim] = data[dim]
@@ -213,17 +220,17 @@ def apply_clustering(data, timesteps, clustering_func, how, normalize=True,
         storage_inter_cluster=storage_inter_cluster
     )
 
-    if timesteps is None:
-        data_new = _copy_non_t_vars(data, data_new)
-    else:
-        # Drop timesteps from old data
+    # It's now safe to add the original coordinates back in (preserving all the
+    # loc_tech sets that aren't used to index a variable in the DataArray)
+    data_new.update(data_coords)
+
+    data_new = _copy_non_t_vars(data, data_new)
+
+    if timesteps is not None:
[…]
