+++ requirements/base.yml
@@ -26,6 +26,6 @@ dependencies:
     - pytest-cov
     - ruamel.yaml<=0.15
     - scikit-learn=0.19.1
-    - xarray=0.10.1
+    - xarray=0.10.8
     - pip:
         - "git+https://github.com/calliope-project/calliope.git"
