### 2018-08-30 3242256
Fix bug & test it
