### 2018-09-03 8584d72
Minor fixes
