### 2018-09-06 52308a2
Convert object coords to string before saving to netcdf
