### 2018-09-10 #825 @jedbrown
I installed the software and followed one of the tutorials -- looks great.

@sjpfenninger Can you fix capitalization in the bib file? I think this is a result of a new JOSS style file, but proper names should be protected: "europe", "great britain", "ecos", "san diego, ca", "south africa", "python". When that is fixed, we'll be ready to archive. Thanks for your patience.
