### 2018-09-11 #825 @sjpfenninger
@jedbrown That should be fixed. I also bumped `codemeta.json` to v0.6.2 (the current version and the version that was online when the review started).
