### 2018-09-11 #825 @jedbrown
@sjpfenninger Looks good. Please archive your repository using Zenodo or similar and report the DOI here.
