### 2018-09-12 #825 @sjpfenninger
The generic Zenodo DOI is 10.5281/zenodo.593292 and the DOI for version 0.6.2 is 10.5281/zenodo.1262406 -- or do you need the specific commit that contains the accepted version of the JOSS paper to be archived?
