### 2018-09-12 #825 @arfon
> @arfon Do we need a fresh archive or is the v0.6.2 DOI sufficient?

Should be sufficient as long as it represents the code associated with this submission (including any changes from the review).
